refactor(auto-reply): split config dispatch pipeline
