+++ src/auto-reply/reply/dispatch-from-config.ts
@@ -1,224 +1,20 @@
 /** Main reply dispatch pipeline from finalized config/context to delivery payloads. */
-import { isParentOwnedBackgroundAcpSession } from "@openclaw/acp-core/session-interaction-mode";
-import {
-  normalizeLowercaseStringOrEmpty,
-  normalizeOptionalString,
-} from "@openclaw/normalization-core/string-coerce";
-import { truncateUtf16Safe } from "@openclaw/normalization-core/utf16-slice";
-import {
-  hasOutboundReplyContent,
-  isFastModeAutoProgressPayload,
-  resolveSendableOutboundReplyParts,
-} from "openclaw/plugin-sdk/reply-payload";
-import { readAcpSessionMeta } from "../../acp/runtime/session-meta.js";
-import {
-  resolveAgentConfig,
-  resolveAgentWorkspaceDir,
-  resolveSessionAgentId,
-} from "../../agents/agent-scope.js";
-import {
-  resolveEffectiveToolPolicy,
-  resolveGroupToolPolicy,
-  resolveInheritedToolPolicyForSession,
-  resolveSubagentToolPolicyForSession,
-} from "../../agents/agent-tools.policy.js";
-import { resolveAgentIdentity } from "../../agents/identity.js";
-import { resolveSessionModelRef } from "../../agents/session-model-ref.js";
-import {
-  isSubagentEnvelopeSession,
-  resolveSubagentCapabilityStore,
-} from "../../agents/subagent-capabilities.js";
-import { isToolAllowedByPolicies } from "../../agents/tool-policy-match.js";
-import { mergeAlsoAllowPolicy, resolveToolProfilePolicy } from "../../agents/tool-policy.js";
-import { isAskUserPromptPending } from "../../agents/tools/ask-user-tool.js";
-import {
-  resolveConversationBindingRecord,
-  touchConversationBindingRecord,
-} from "../../bindings/records.js";
-import { normalizeChatType } from "../../channels/chat-type.js";
-import { shouldSuppressLocalExecApprovalPrompt } from "../../channels/plugins/exec-approval-local.js";
-import {
-  type AgentPlanStep,
-  formatPlanChecklistLines,
-  normalizeAgentPlanSteps,
-} from "../../channels/streaming.js";
-import { getRuntimeConfigSnapshot } from "../../config/config.js";
-import { applyMergePatch } from "../../config/merge-patch.js";
-import { normalizeExplicitSessionKey } from "../../config/sessions/explicit-session-key-normalization.js";
-import { resolveGroupSessionKey } from "../../config/sessions/group.js";
-import type { OpenClawConfig } from "../../config/types.openclaw.js";
-import { logVerbose } from "../../globals.js";
-import { fireAndForgetHook } from "../../hooks/fire-and-forget.js";
-import {
-  deriveInboundMessageHookContext,
-  toPluginInboundClaimContext,
-  toPluginInboundClaimEvent,
-  toInternalMessageReceivedContext,
-  toPluginMessageContext,
-  toPluginMessageReceivedEvent,
-} from "../../hooks/message-hook-mappers.js";
-import { isDiagnosticsEnabled } from "../../infra/diagnostic-events.js";
-import { measureDiagnosticsTimelineSpan } from "../../infra/diagnostics-timeline.js";
-import { formatErrorMessage } from "../../infra/errors.js";
-import {
-  logMessageDispatchCompleted,
-  logMessageDispatchStarted,
-  markDiagnosticSessionProgress,
-} from "../../logging/diagnostic.js";
-import { createDiagnosticMessageLifecycle } from "../../logging/message-lifecycle.js";
-import {
-  buildPluginBindingDeclinedText,
-  buildPluginBindingErrorText,
-  buildPluginBindingUnavailableText,
-  hasShownPluginBindingFallbackNotice,
-  isPluginOwnedSessionBindingRecord,
-  markPluginBindingFallbackNoticeShown,
-  toPluginConversationBinding,
-} from "../../plugins/conversation-binding.js";
-import { getGlobalHookRunner, getGlobalPluginRegistry } from "../../plugins/hook-runner-global.js";
-import type { PluginHookReplyDispatchEvent } from "../../plugins/hook-types.js";
-import { resolveSendPolicy } from "../../sessions/send-policy.js";
-import { resolveSilentReplyPolicyFromPolicies } from "../../shared/silent-reply-policy.js";
-import { createTtsDirectiveTextStreamCleaner } from "../../tts/directives.js";
-import {
-  normalizeTtsAutoMode,
-  resolveConfiguredTtsMode,
-  shouldCleanTtsDirectiveText,
-} from "../../tts/tts-config.js";
-import { INTERNAL_MESSAGE_CHANNEL, normalizeMessageChannel } from "../../utils/message-channel.js";
-import { resolveCommandAuthorization } from "../command-auth.js";
-import { resolveCommandTurnTargetSessionKey } from "../command-turn-context.js";
-import { registerReplyDispatcherSettledTask } from "../dispatch-dispatcher.js";
-import type { BlockReplyContext, GetReplyOptions } from "../get-reply-options.types.js";
-import {
-  copyReplyPayloadMetadata,
-  getReplyPayloadMetadata,
-  isReplyPayloadStatusNotice,
-  markReplyPayloadAsTtsSupplement,
-  setReplyPayloadMetadata,
-  type ReplyPayload,
-} from "../reply-payload.js";
-import type { FinalizedRuntimeMsgContext as FinalizedMsgContext } from "../templating.js";
-import { normalizeVerboseLevel } from "../thinking.js";
-import {
-  takeCommandSessionMetadataChanges,
-  type CommandSessionMetadataChange,
-} from "./command-session-metadata.js";
-import { resolveConversationBindingContextFromMessage } from "./conversation-binding-input.js";
-import { capturePendingConversationTurnReply } from "./conversation-turn-capture.js";
-import {
-  DispatchReplyOperationAbortedError,
-  isDispatchReplyOperationAbortedError,
-  runWithDispatchAbortSignal,
-} from "./dispatch-from-config.abort.js";
-import {
-  createInboundMessageAuditTerminal,
-  type DispatchProcessedOptions,
-  type DispatchProcessedOutcome,
-  type InboundMessageAuditTerminalRecorder,
-} from "./dispatch-from-config.audit.js";
-import {
-  resolveBoundAcpDispatchSessionKey,
-  resolveRoutedPolicyConversationType,
-  resolveSessionStoreLookup,
-} from "./dispatch-from-config.context.js";
-import {
-  createShouldEmitVerboseProgress,
-  resolveHarnessSourceVisibleRepliesDefault,
-  resolveTurnModelOverride,
-} from "./dispatch-from-config.harness-defaults.js";
-import { createDispatchReplyOperationCoordinator } from "./dispatch-from-config.lifecycle.js";
-import {
-  createFinalizationAwareTtsPayloadApplier,
-  createFinalDispatchPayloadDedupeKey,
-  formatSuppressedReplyPayloadForLog,
-} from "./dispatch-from-config.payloads.js";
-import {
-  clearPendingFinalDeliveryAfterSuccess,
-  capturePendingFinalDeliveryIdentity,
-  reconcilePendingFinalDeliveryAfterSettlement,
-} from "./dispatch-from-config.pending-final.js";
-import { shouldBypassPluginOwnedBindingForCommand } from "./dispatch-from-config.plugin-binding.js";
-import {
-  loadAbortRuntime,
-  loadGetReplyFromConfigRuntime,
-  loadReplyMediaPathsRuntime,
-  loadRouteReplyRuntime,
-  loadRuntimePlugins,
-} from "./dispatch-from-config.runtime-loaders.js";
-import { createInternalHookEvent, triggerInternalHook } from "./dispatch-from-config.runtime.js";
-import { createReplyHotPathTimingTracker } from "./dispatch-from-config.timing.js";
-import {
-  captureDeliveredTranscriptMirror,
-  getDispatcherFinalOutcomeCounts,
-  mirrorDeliveredReplyToTranscript,
-  mirrorTranscriptAfterDispatcherSettled,
-  transcriptMirrorForDeliveredPayload,
-} from "./dispatch-from-config.transcript.js";
+import { isDispatchReplyOperationAbortedError } from "./dispatch-from-config.abort.js";
+import { createInboundMessageAuditTerminal } from "./dispatch-from-config.audit.js";
+import { chooseDispatchRoute } from "./dispatch-from-config.choose-route.js";
+import { executeDispatch } from "./dispatch-from-config.execute.js";
+import { finalizeDispatchAndAudit } from "./dispatch-from-config.finalize.js";
+import { gatherDispatchRequest } from "./dispatch-from-config.gather.js";
+import { prepareDispatchOperationContext } from "./dispatch-from-config.prepare-context.js";
+import { prepareDispatchDelivery } from "./dispatch-from-config.prepare-delivery.js";
+import { prepareDispatchExecution } from "./dispatch-from-config.prepare-execution.js";
+import { prepareDispatchOperation } from "./dispatch-from-config.prepare-operation.js";
 import type {
   DispatchFromConfigParams,
   DispatchFromConfigResult,
 } from "./dispatch-from-config.types.js";
-import { resolveEffectiveReplyRoute } from "./effective-reply-route.js";
-import { withFullRuntimeReplyConfig } from "./get-reply-fast-path.js";
-import type { ReplySessionBinding } from "./get-reply.types.js";
-import { stripLegacyMediaContextFields } from "./inbound-context.js";
-import { claimInboundDedupe, commitInboundDedupe, releaseInboundDedupe } from "./inbound-dedupe.js";
-import { hasInboundAudio } from "./inbound-media.js";
-import { resolveOriginMessageProvider } from "./origin-routing.js";
-import {
-  captureReplyDispatchDeliveryOutcome,
-  type ReplyDispatchDeliveryOutcome,
-  waitForReplyDispatcherIdle,
-} from "./reply-dispatcher.js";
-import type { ReplyDispatchKind } from "./reply-dispatcher.types.js";
-import { replyRunRegistry } from "./reply-run-registry.js";
-import {
-  createReplyDeliveryContext,
-  resolveReplyDeliveryAccountId,
-  resolveReplyToMode,
-} from "./reply-threading.js";
-import { isReplyProfilerEnabled } from "./reply-timing-tracker.js";
-import { extractShortModelName, type ResponsePrefixContext } from "./response-prefix-template.js";
-import { isDuplicateRestartRecoverySource } from "./restart-recovery-claim.js";
-import { resolveRoutedDeliveryThreadId } from "./routed-delivery-thread.js";
-import { resolveReplyRoutingDecision } from "./routing-policy.js";
-import {
-  isExplicitSourceReplyCommand,
-  isUnauthorizedTextSlashCommand,
-  resolveSourceReplyVisibilityPolicy,
-} from "./source-reply-delivery-mode.js";
-import {
-  buildChannelSourceTurnId,
-  readChannelSourceTurnId,
-  setChannelSourceTurnId,
-  shouldMintChannelSourceTurnId,
-} from "./source-turn-id.js";
-import { stageRemoteInboundMediaIfNeeded } from "./stage-remote-inbound-media.js";
-import { resolveRunTypingPolicy } from "./typing-policy.js";
-
-type InternalReplyResolverOptions = {
-  onSessionMetadataChanges?: (changes: CommandSessionMetadataChange[]) => void;
-  onSessionPrepared?: (binding: ReplySessionBinding) => void;
-};
-
-function createReplyDispatchEvent(
-  params: Omit<PluginHookReplyDispatchEvent, "shouldSendToolSummaries"> & {
-    shouldSendToolSummaries: () => boolean;
-  },
-): PluginHookReplyDispatchEvent {
-  const { shouldSendToolSummaries, ...event } = params;
-  return Object.defineProperty(event, "shouldSendToolSummaries", {
-    enumerable: true,
-    get: shouldSendToolSummaries,
-  }) as PluginHookReplyDispatchEvent;
-}
-
-if (process.env.VITEST || process.env.NODE_ENV === "test") {
-  (globalThis as Record<PropertyKey, unknown>)[Symbol.for("openclaw.dispatchFromConfigTestApi")] = {
-    createReplyDispatchEvent,
-  };
-}
+import { commitInboundDedupe, releaseInboundDedupe } from "./inbound-dedupe.js";
+import "./dispatch-from-config.events.js";
 
 export type { DispatchFromConfigResult } from "./dispatch-from-config.types.js";
 
@@ -239,2813 +35,55 @@ export async function dispatchReplyFromConfig(
 
 async function dispatchReplyFromConfigInner(
   params: DispatchFromConfigParams,
-  messageAuditTerminal: InboundMessageAuditTerminalRecorder | undefined,
+  messageAuditTerminal: ReturnType<typeof createInboundMessageAuditTerminal>,
 ): Promise<DispatchFromConfigResult> {
-  const { ctx, cfg, dispatcher } = params;
-  if (params.replyOptions?.abortSignal?.aborted) {
-    messageAuditTerminal?.note("skipped", { reason: "reply_operation_aborted" });
-    return {
-      queuedFinal: false,
-      counts: dispatcher.getQueuedCounts(),
-    };
-  }
-  const diagnosticsEnabled = isDiagnosticsEnabled(cfg);
-  const channel = normalizeLowercaseStringOrEmpty(ctx.Surface ?? ctx.Provider ?? "unknown");
-  const chatId = ctx.To ?? ctx.From;
-  const messageId =
-    ctx.MessageSidFull ?? ctx.MessageSid ?? ctx.MessageSidFirst ?? ctx.MessageSidLast;
-  const sessionKey =
-    normalizeOptionalString(ctx.SessionKey) ?? normalizeOptionalString(ctx.CommandTargetSessionKey);
-  const startTime = diagnosticsEnabled ? Date.now() : 0;
-  const canTrackSession = diagnosticsEnabled && Boolean(sessionKey);
-  const initialSessionStoreEntry = resolveSessionStoreLookup(ctx, cfg);
-  // resolveSessionStoreLookup is command-target-aware (it prefers
-  // resolveCommandTurnTargetSessionKey), whereas the lifecycle's sessionKey is
-  // source-first (ctx.SessionKey). On a native command turn that targets a
-  // different session, the resolved entry can belong to the *target* while the
-  // lifecycle reports the *source* key — so only carry the UUID when the entry
-  // is for the same session the lifecycle reports, to avoid mis-associating a
-  // session id with the wrong session key. When they diverge, emit sessionKey
-  // only (prior behavior).
-  const lifecycleSessionId =
-    initialSessionStoreEntry.sessionKey === sessionKey
-      ? initialSessionStoreEntry.entry?.sessionId
-      : undefined;
-  const messageLifecycle = createDiagnosticMessageLifecycle({
-    enabled: diagnosticsEnabled,
-    channel,
-    chatId,
-    messageId,
-    sessionKey,
-    sessionId: lifecycleSessionId,
-    source: "dispatch",
-    processingReason: "message_start",
-    startedAtMs: startTime,
-    trackSessionState: canTrackSession,
-  });
-  const traceAttributes = {
-    surface: channel,
-    hasSessionKey: Boolean(sessionKey),
-    hasRunId: typeof params.replyOptions?.runId === "string",
-  };
-  const replyHotPathTiming = createReplyHotPathTimingTracker({
-    profilerEnabled: isReplyProfilerEnabled({ config: cfg }),
-  });
-  const traceReplyPhase = <T>(name: string, run: () => Promise<T> | T): Promise<T> =>
-    replyHotPathTiming.measure(name, () =>
-      measureDiagnosticsTimelineSpan(name, run, {
-        phase: "agent-turn",
-        config: cfg,
-        attributes: traceAttributes,
-      }),
-    );
-  let agentDispatchStartedAt = 0;
-
-  const recordProcessed = (outcome: DispatchProcessedOutcome, opts?: DispatchProcessedOptions) => {
-    messageAuditTerminal?.note(outcome, opts);
-    if (diagnosticsEnabled) {
-      replyHotPathTiming.logIfSlow({
-        channel,
-        messageId,
-        sessionKey,
-        outcome,
-        reason: opts?.reason,
-      });
-    }
-    messageLifecycle.markProcessed(outcome, opts);
-  };
-
-  const recordAgentDispatchStarted = () => {
-    if (!diagnosticsEnabled || agentDispatchStartedAt > 0) {
-      return;
-    }
-    agentDispatchStartedAt = Date.now();
-    logMessageDispatchStarted({
-      channel,
-      sessionKey: acpDispatchSessionKey,
-      source: "replyResolver",
-    });
-  };
-
-  const recordAgentDispatchCompleted = (
-    outcome: "completed" | "skipped" | "error",
-    opts?: {
-      reason?: string;
-      error?: string;
-    },
-  ) => {
-    if (!diagnosticsEnabled || agentDispatchStartedAt <= 0) {
-      return;
-    }
-    logMessageDispatchCompleted({
-      channel,
-      sessionKey: acpDispatchSessionKey,
-      source: "replyResolver",
-      durationMs: Date.now() - agentDispatchStartedAt,
-      outcome,
-      reason: opts?.reason,
-      error: opts?.error,
-    });
-  };
-
-  const markProcessing = () => {
-    messageLifecycle.markProcessing();
-  };
-
-  const markIdle = (reason: string) => {
-    messageLifecycle.markIdle(reason);
-  };
-
-  let inboundDedupeReplayUnsafe = false;
-  const markInboundDedupeReplayUnsafe = () => {
-    inboundDedupeReplayUnsafe = true;
-  };
-
-  const boundAcpDispatchSessionKey = resolveBoundAcpDispatchSessionKey({ ctx, cfg });
-  const acpDispatchSessionKey =
-    boundAcpDispatchSessionKey ?? initialSessionStoreEntry.sessionKey ?? sessionKey;
-  // initialSessionStoreEntry stays command-target-aware for handler/store
-  // lookups (status/stop/model act on the target via CommandTargetSessionKey).
-  // Reply-run ownership must stay SOURCE-keyed: a native command turn must not
-  // wait on or contend with the target's active run. Bound ACP routing uses
-  // acpDispatchSessionKey separately and must not move source admission.
-  const sourceSessionKey = normalizeOptionalString(ctx.SessionKey);
-  const dispatchOperationSessionKey =
-    sourceSessionKey ?? initialSessionStoreEntry.sessionKey ?? sessionKey ?? acpDispatchSessionKey;
-  const operationSessionStoreEntry =
-    sourceSessionKey &&
-    initialSessionStoreEntry.sessionKey &&
-    sourceSessionKey !== initialSessionStoreEntry.sessionKey
-      ? resolveSessionStoreLookup(
-          {
-            ...ctx,
-            // Strip target so store resolution follows the source SessionKey.
-            CommandTargetSessionKey: undefined,
-          },
-          cfg,
-        )
-      : initialSessionStoreEntry;
-  const initialDispatchReplyOperation = dispatchOperationSessionKey
-    ? replyRunRegistry.get(dispatchOperationSessionKey)
-    : undefined;
-  if (
-    params.replyOptions?.isHeartbeat === true &&
-    dispatchOperationSessionKey &&
-    initialDispatchReplyOperation
-  ) {
-    messageAuditTerminal?.note("skipped", { reason: "reply-operation-active" });
-    return {
-      queuedFinal: false,
-      counts: dispatcher.getQueuedCounts(),
-    };
-  }
-  const markProgress = () => {
-    if (!canTrackSession || !sessionKey) {
-      return;
-    }
-    markDiagnosticSessionProgress({ sessionKey });
-    if (acpDispatchSessionKey && acpDispatchSessionKey !== sessionKey) {
-      markDiagnosticSessionProgress({ sessionKey: acpDispatchSessionKey });
-    }
-  };
-  const sessionStoreEntry = boundAcpDispatchSessionKey
-    ? resolveSessionStoreLookup({ ...ctx, SessionKey: boundAcpDispatchSessionKey }, cfg)
-    : initialSessionStoreEntry;
-  let preparedSessionBinding: ReplySessionBinding | undefined =
-    sessionStoreEntry.sessionKey && sessionStoreEntry.entry?.sessionId
-      ? {
-          sessionKey: sessionStoreEntry.sessionKey,
-          sessionId: sessionStoreEntry.entry.sessionId,
-          storePath: sessionStoreEntry.storePath,
-        }
-      : undefined;
-  let preparedOperationSessionBinding: ReplySessionBinding | undefined =
-    operationSessionStoreEntry.sessionKey && operationSessionStoreEntry.entry?.sessionId
-      ? {
-          sessionKey: operationSessionStoreEntry.sessionKey,
-          sessionId: operationSessionStoreEntry.entry.sessionId,
-          storePath: operationSessionStoreEntry.storePath,
-        }
-      : undefined;
-  const sessionKeysMatch = (left?: string, right?: string) =>
-    Boolean(
-      left &&
-      right &&
-      normalizeExplicitSessionKey(left, ctx) === normalizeExplicitSessionKey(right, ctx),
-    );
-  const notePreparedSession = (binding: ReplySessionBinding) => {
-    if (sessionKeysMatch(binding.sessionKey, sessionStoreEntry.sessionKey)) {
-      preparedSessionBinding = binding;
-    }
-    if (sessionKeysMatch(binding.sessionKey, operationSessionStoreEntry.sessionKey)) {
-      preparedOperationSessionBinding = binding;
-    }
-    params.replyOptions?.onSessionPrepared?.(binding);
-  };
-  const resolveOperationExpectedSessionId = () =>
-    preparedOperationSessionBinding?.sessionId ?? operationSessionStoreEntry.entry?.sessionId;
-  const resolvePreparedTranscriptBinding = (mirrorSessionKey?: string) => {
-    if (
-      !preparedSessionBinding ||
-      !sessionKeysMatch(mirrorSessionKey, preparedSessionBinding.sessionKey)
-    ) {
-      return undefined;
-    }
-    return preparedSessionBinding;
-  };
-  const sessionAgentId = resolveSessionAgentId({
-    sessionKey: acpDispatchSessionKey,
-    config: cfg,
-    fallbackAgentId: ctx.AgentId,
-  });
-  const sessionAgentCfg = resolveAgentConfig(cfg, sessionAgentId);
-  const verboseProgress = createShouldEmitVerboseProgress({
-    agentId: sessionAgentId,
-    sessionKey: acpDispatchSessionKey,
-    storePath: sessionStoreEntry.storePath,
-    initialExplicitLevel: sessionStoreEntry.entry?.verboseLevel,
-    fallbackLevel:
-      normalizeVerboseLevel(
-        sessionStoreEntry.entry?.verboseLevel ??
-          sessionAgentCfg?.verboseDefault ??
-          cfg.agents?.defaults?.verboseDefault ??
-          "",
-      ) ?? "off",
-  });
-  const shouldEmitVerboseProgress = verboseProgress.shouldEmit;
-  const shouldEmitFullVerboseProgress = verboseProgress.shouldEmitFull;
-  const replyRoute = resolveEffectiveReplyRoute({ ctx, entry: sessionStoreEntry.entry });
-  // Restore route thread context only from the active turn or the thread-scoped session key.
-  // Do not read thread ids from the normalised session store here: `origin.threadId` can be
-  // folded back into lastThreadId/deliveryContext during store normalisation and resurrect a
-  // stale route after thread delivery was intentionally cleared.
-  const routeThreadId = resolveRoutedDeliveryThreadId({
-    ctx,
-    sessionKey: acpDispatchSessionKey,
-  });
-  // Inherited sessions_send routes carry thread ids only when the stored route
-  // proves the thread came from an explicit target, not session normalization.
-  const routeReplyThreadId = replyRoute.threadId ?? routeThreadId;
-  const inboundAudio = hasInboundAudio(ctx);
-  const sessionTtsAuto = normalizeTtsAutoMode(sessionStoreEntry.entry?.ttsAuto);
-  const workspaceDir = resolveAgentWorkspaceDir(cfg, sessionAgentId);
-  const replyOperationCoordinator = createDispatchReplyOperationCoordinator({
-    ctx,
-    dispatcher,
-    dispatchOperationSessionKey,
-    initialDispatchReplyOperation,
-    messageAuditTerminal,
-    operationSessionStoreEntry,
-    replyOptions: params.replyOptions,
-    resolveOperationExpectedSessionId,
-    routeThreadId,
-  });
-  const {
-    completeDispatchReplyOperation,
-    dispatchHookDispatcher,
-    ensureDispatchReplyOperation,
-    failDispatchReplyOperation,
-    getDispatchAbortOperation,
-    getDispatchAbortSignal,
-    getDispatchReplyOperation,
-    getObservedReplyDelivery,
-    getPreDispatchAbortSignal,
-    getReplyOptions,
-    isDispatchOperationAborted,
-    isPreDispatchOperationAborted,
-    markObservedReplyDelivery,
-    releasePreDispatchLifecycleAdmission,
-    runWithDispatchLifecycleAdmission,
-    throwIfDispatchOperationAborted,
-    trackDispatchLifecycleWork,
-  } = replyOperationCoordinator;
-  const maybeApplyTtsWithFinalizationLease = createFinalizationAwareTtsPayloadApplier({
-    getReplyOperation: getDispatchReplyOperation,
-    hasInboundAudio: () =>
-      inboundAudio || getDispatchReplyOperation()?.acceptedSteeredInboundAudio === true,
-  });
-  const { ensureRuntimePluginsLoaded } = await traceReplyPhase("reply.load_runtime_plugins", () =>
-    loadRuntimePlugins(),
-  );
-  await traceReplyPhase("reply.ensure_runtime_plugins", () => {
-    ensureRuntimePluginsLoaded({ config: cfg, workspaceDir });
-  });
-  const hookRunner = getGlobalHookRunner();
-  // Extract message context for hooks (plugin and internal)
-  const timestamp =
-    typeof ctx.Timestamp === "number" && Number.isFinite(ctx.Timestamp) ? ctx.Timestamp : undefined;
-  const messageIdForHook =
-    ctx.MessageSidFull ?? ctx.MessageSid ?? ctx.MessageSidFirst ?? ctx.MessageSidLast;
-  const hookCtx = { ...ctx };
-  const buildHookState = (sourceCtx: FinalizedMsgContext) => {
-    const nextHookContext = deriveInboundMessageHookContext(sourceCtx, {
-      messageId: messageIdForHook,
-    });
-    return {
-      hookContext: nextHookContext,
-      inboundClaimContext: toPluginInboundClaimContext(nextHookContext),
-      inboundClaimEvent: toPluginInboundClaimEvent(nextHookContext, {
-        commandAuthorized:
-          typeof ctx.CommandAuthorized === "boolean" ? ctx.CommandAuthorized : undefined,
-        wasMentioned: typeof ctx.WasMentioned === "boolean" ? ctx.WasMentioned : undefined,
-      }),
-    };
-  };
-  let { hookContext, inboundClaimContext, inboundClaimEvent } = buildHookState(hookCtx);
-  const { isGroup, groupId } = hookContext;
-  let hookMediaPrepared = false;
-  let hookMediaMetadataStaged = false;
-  const prepareHookMediaMetadata = async () => {
-    if (hookMediaPrepared) {
-      return;
-    }
-    hookMediaPrepared = true;
-    // Plugin hooks may run in a different Codex cwd from core dispatch, so
-    // only actual hook/plugin-claim consumers get remote-cache media paths.
-    // Keep ctx unstaged for the normal get-reply single-stage path.
-    const staged = await traceReplyPhase("reply.stage_remote_media_for_dispatch", () =>
-      stageRemoteInboundMediaIfNeeded({
-        ctx: hookCtx,
-        cfg,
-        sessionKey: acpDispatchSessionKey,
-        workspaceDir,
-        remoteMediaMode: "cache",
-      }),
-    );
-    if (staged) {
-      hookMediaMetadataStaged = true;
-      ({ hookContext, inboundClaimContext, inboundClaimEvent } = buildHookState(hookCtx));
-    }
-  };
-  const buildMessageReceivedHookContext = () => {
-    const mediaRemoteHost = normalizeOptionalString(ctx.MediaRemoteHost);
-    const hasUnstagedRemoteMediaMetadata = Boolean(
-      hookContext.mediaPath ||
-      hookContext.mediaUrl ||
-      hookContext.mediaType ||
-      hookContext.mediaPaths?.length ||
-      hookContext.mediaUrls?.length ||
-      hookContext.mediaTypes?.length,
-    );
-    if (hookMediaMetadataStaged || !mediaRemoteHost || !hasUnstagedRemoteMediaMetadata) {
-      return hookContext;
-    }
-    const messageReceivedCtx = { ...hookCtx };
-    // message_received hooks run before normal get-reply staging, so remote
-    // host paths are not safe as live media. Keep originals as debug metadata.
-    stripLegacyMediaContextFields(messageReceivedCtx);
-    delete messageReceivedCtx.media;
-    return {
-      ...buildHookState(messageReceivedCtx).hookContext,
-      mediaRemoteHost,
-      mediaStagingPending: true,
-      originalMediaPath: hookContext.mediaPath,
-      originalMediaUrl: hookContext.mediaUrl,
-      originalMediaType: hookContext.mediaType,
-      originalMediaPaths: hookContext.mediaPaths,
-      originalMediaUrls: hookContext.mediaUrls,
-      originalMediaTypes: hookContext.mediaTypes,
-    };
-  };
-
-  // Check if we should route replies to originating channel instead of dispatcher.
-  // Only route when the originating channel is DIFFERENT from the current surface.
-  // This handles cross-provider routing (e.g., message from Telegram being processed
-  // by a shared session that's currently on Slack) while preserving normal dispatcher
-  // flow when the provider handles its own messages.
-  //
-  // Debug: `pnpm test src/auto-reply/reply/dispatch-from-config.test.ts`
-  const sessionAcpMeta = sessionStoreEntry.sessionKey
-    ? readAcpSessionMeta({ sessionKey: sessionStoreEntry.sessionKey })
-    : undefined;
-  const sessionEntryWithAcp =
-    sessionAcpMeta && sessionStoreEntry.entry
-      ? { ...sessionStoreEntry.entry, acp: sessionAcpMeta }
-      : sessionStoreEntry.entry;
-  const suppressAcpChildUserDelivery = isParentOwnedBackgroundAcpSession(sessionEntryWithAcp);
-  const normalizedRouteReplyChannel = normalizeMessageChannel(replyRoute.channel);
-  const normalizedProviderChannel = normalizeMessageChannel(ctx.Provider);
-  const normalizedSurfaceChannel = normalizeMessageChannel(ctx.Surface);
-  const normalizedCurrentSurface = normalizedProviderChannel ?? normalizedSurfaceChannel;
-  const effectiveExplicitDeliverRoute =
-    ctx.ExplicitDeliverRoute === true || replyRoute.inheritedExternalRoute === true;
-  const isInternalWebchatTurn =
-    normalizedCurrentSurface === INTERNAL_MESSAGE_CHANNEL &&
-    (normalizedSurfaceChannel === INTERNAL_MESSAGE_CHANNEL || !normalizedSurfaceChannel) &&
-    !effectiveExplicitDeliverRoute;
-  const hasRouteReplyCandidate = Boolean(
-    !suppressAcpChildUserDelivery &&
-    !isInternalWebchatTurn &&
-    normalizedRouteReplyChannel &&
-    replyRoute.to &&
-    normalizedRouteReplyChannel !== normalizedCurrentSurface,
-  );
-  const routeReplyRuntime = hasRouteReplyCandidate ? await loadRouteReplyRuntime() : undefined;
-  const {
-    originatingChannel: routeReplyChannel,
-    currentSurface,
-    shouldRouteToOriginating,
-    shouldSuppressTyping,
-  } = resolveReplyRoutingDecision({
-    provider: ctx.Provider,
-    surface: ctx.Surface,
-    explicitDeliverRoute: effectiveExplicitDeliverRoute,
-    originatingChannel: replyRoute.channel,
-    originatingTo: replyRoute.to,
-    suppressDirectUserDelivery: suppressAcpChildUserDelivery,
-    isRoutableChannel: routeReplyRuntime?.isRoutableChannel ?? (() => false),
-  });
-  const routeReplyTo = replyRoute.to;
-  const deliveryChannel = shouldRouteToOriginating ? routeReplyChannel : currentSurface;
-  const shouldPrepareRoutedReplyDelivery = shouldRouteToOriginating && Boolean(routeReplyChannel);
-  const replyContextAccountId = routeReplyChannel
-    ? resolveReplyDeliveryAccountId(cfg, routeReplyChannel, replyRoute.accountId)
-    : undefined;
-  const routedReplyAccountId = shouldPrepareRoutedReplyDelivery ? replyContextAccountId : undefined;
-  const routedReplyDelivery = shouldPrepareRoutedReplyDelivery
-    ? createReplyDeliveryContext(
-        resolveReplyToMode(cfg, routeReplyChannel, routedReplyAccountId, replyRoute.chatType),
-        replyRoute.chatType,
-      )
-    : undefined;
-  let normalizeReplyMediaPaths:
-    | ReturnType<
-        (typeof import("./reply-media-paths.runtime.js"))["createReplyMediaPathNormalizer"]
-      >
-    | undefined;
-  const getNormalizeReplyMediaPaths = async () => {
-    if (normalizeReplyMediaPaths) {
-      return normalizeReplyMediaPaths;
-    }
-    const { createReplyMediaPathNormalizer } = await loadReplyMediaPathsRuntime();
-    normalizeReplyMediaPaths = createReplyMediaPathNormalizer({
-      cfg,
-      sessionKey: acpDispatchSessionKey,
-      workspaceDir,
-      messageProvider: deliveryChannel,
-      accountId: replyContextAccountId,
-      groupId,
-      groupChannel: ctx.GroupChannel,
-      groupSpace: ctx.GroupSpace,
-      requesterSenderId: ctx.SenderId,
-      requesterSenderName: ctx.SenderName,
-      requesterSenderUsername: ctx.SenderUsername,
-      requesterSenderE164: ctx.SenderE164,
-    });
-    return normalizeReplyMediaPaths;
-  };
-  const normalizeReplyMediaPayload = async (payload: ReplyPayload): Promise<ReplyPayload> => {
-    if (!resolveSendableOutboundReplyParts(payload).hasMedia) {
-      return payload;
-    }
-    const normalizeReplyMediaPayloadPaths = await getNormalizeReplyMediaPaths();
-    return await normalizeReplyMediaPayloadPaths(payload);
-  };
-
-  const routeReplyToOriginating = async (
-    payload: ReplyPayload,
-    options?: {
-      abortSignal?: AbortSignal;
-      mirror?: boolean;
-      kind?: ReplyDispatchKind;
-      responsePrefixContext?: ResponsePrefixContext;
-      sessionKey?: string;
-    },
-  ) => {
-    if (!shouldRouteToOriginating || !routeReplyChannel || !routeReplyTo || !routeReplyRuntime) {
-      return null;
-    }
-    markInboundDedupeReplayUnsafe();
-    // Outbound session.key must match the session key used by the agent
-    // runtime that produced this payload, so agent_end and message delivery
-    // hooks expose the same canonical key for native command redirects.
-    const agentRuntimeSessionKey =
-      options?.sessionKey ??
-      (ctx.CommandSource === "native"
-        ? (resolveCommandTurnTargetSessionKey(ctx) ?? ctx.SessionKey)
-        : ctx.SessionKey);
-    return await routeReplyRuntime.routeReply({
-      payload,
-      channel: routeReplyChannel,
-      to: routeReplyTo,
-      sessionKey: agentRuntimeSessionKey,
-      policySessionKey:
-        options?.sessionKey ?? resolveCommandTurnTargetSessionKey(ctx) ?? ctx.SessionKey,
-      policyConversationType: resolveRoutedPolicyConversationType(ctx),
-      accountId: routedReplyAccountId,
-      requesterSenderId: ctx.SenderId,
-      requesterSenderName: ctx.SenderName,
-      requesterSenderUsername: ctx.SenderUsername,
-      requesterSenderE164: ctx.SenderE164,
-      threadId: routeReplyThreadId,
-      replyDelivery: routedReplyDelivery,
-      cfg,
-      abortSignal: options?.abortSignal,
-      mirror: options?.mirror,
-      isGroup,
-      groupId,
-      replyKind: options?.kind ?? "final",
-      runId: params.replyOptions?.runId,
-      responsePrefixContext: options?.responsePrefixContext,
-    });
-  };
-
-  const isRoutedReplyDelivered = (result: { ok: boolean; suppressed?: boolean }) =>
-    result.ok && result.suppressed !== true;
-
-  /**
-   * Helper to send a payload via route-reply (async).
-   * Only used when actually routing to a different provider.
-   * Note: Only called when shouldRouteToOriginating is true, so
-   * routeReplyChannel and routeReplyTo are guaranteed to be defined.
-   */
-  const sendPayloadAsync = async (
-    payload: ReplyPayload,
-    abortSignal?: AbortSignal,
-    mirror?: boolean,
-    kind: ReplyDispatchKind = "tool",
-  ): Promise<void> => {
-    // Keep the runtime guard explicit because this helper is called from nested
-    // reply callbacks where TypeScript cannot narrow shouldRouteToOriginating.
-    if (!routeReplyRuntime || !routeReplyChannel || !routeReplyTo) {
-      return;
-    }
-    const effectiveAbortSignal = abortSignal ?? getDispatchAbortSignal();
-    if (effectiveAbortSignal?.aborted) {
-      return;
-    }
-    const result = await routeReplyToOriginating(payload, {
-      abortSignal: effectiveAbortSignal,
-      mirror,
-      kind,
-    });
-    if (result && !result.ok) {
-      logVerbose(`dispatch-from-config: route-reply failed: ${result.error ?? "unknown error"}`);
-    }
-  };
-
-  type PluginBindingTranscriptOwner = {
-    agentId: string;
-    expectedSessionId?: string;
-    sessionKey: string;
-    transcriptWriteBlocked?: true;
-  };
-  const deliverBindingPayload = async (
-    payload: ReplyPayload,
-    mode: "additive" | "terminal",
-    transcriptOwner?: PluginBindingTranscriptOwner,
-  ): Promise<boolean> => {
-    // Metadata is delivery-specific. Keep it off the plugin-owned payload so a
-    // reused reply object cannot carry a stale transcript owner into a later turn.
-    const bindingPayload = setReplyPayloadMetadata(
-      copyReplyPayloadMetadata(payload, { ...payload }),
-      {
-        sourceReplyTranscriptMirror: transcriptOwner
-          ? {
-              sessionKey: transcriptOwner.sessionKey,
-              agentId: transcriptOwner.agentId,
-              ...(transcriptOwner.expectedSessionId
-                ? { expectedSessionId: transcriptOwner.expectedSessionId }
-                : {}),
-              ...(transcriptOwner.transcriptWriteBlocked ? { transcriptWriteBlocked: true } : {}),
-            }
-          : undefined,
-      },
-    );
-    const result = await routeReplyToOriginating(bindingPayload, {
-      kind: mode === "terminal" ? "final" : "tool",
-      sessionKey: transcriptOwner?.sessionKey,
-    });
-    if (result) {
-      if (!result.ok) {
-        logVerbose(
-          `dispatch-from-config: route-reply (plugin binding notice) failed: ${result.error ?? "unknown error"}`,
-        );
-      }
-      return result.ok;
-    }
-    markInboundDedupeReplayUnsafe();
-    return mode === "additive"
-      ? dispatcher.sendToolResult(bindingPayload)
-      : dispatcher.sendFinalReply(bindingPayload);
-  };
-  const sendBindingNotice = async (
-    payload: ReplyPayload,
-    mode: "additive" | "terminal",
-    transcriptOwner?: PluginBindingTranscriptOwner,
-  ): Promise<boolean> => {
-    if (suppressAutomaticSourceDelivery) {
-      return false;
-    }
-    return await deliverBindingPayload(payload, mode, transcriptOwner);
-  };
-
-  // Hook contexts use transport-native ids (for example Slack `U123`), while
-  // binding records use the channel's canonical target (`user:U123`). Resolve
-  // through the binding contract instead of reusing the hook projection.
-  const pluginBindingConversation = resolveConversationBindingContextFromMessage({ cfg, ctx });
-  const pluginOwnedBindingRecord = pluginBindingConversation
-    ? resolveConversationBindingRecord({
-        channel: pluginBindingConversation.channel,
-        accountId: pluginBindingConversation.accountId,
-        conversationId: pluginBindingConversation.conversationId,
-        parentConversationId: pluginBindingConversation.parentConversationId,
-      })
-    : null;
-  const pluginOwnedBinding = isPluginOwnedSessionBindingRecord(pluginOwnedBindingRecord)
-    ? toPluginConversationBinding(pluginOwnedBindingRecord)
-    : null;
-  const pluginBindingSessionKey = normalizeOptionalString(
-    pluginOwnedBindingRecord?.targetSessionKey,
-  );
-  const persistPluginBindingUserTurn = async (): Promise<
-    PluginBindingTranscriptOwner | undefined
-  > => {
-    const recorder = params.replyOptions?.userTurnTranscriptRecorder;
-    if (!recorder || !pluginBindingSessionKey) {
-      return undefined;
-    }
-    const targetAgentId = resolveSessionAgentId({
-      sessionKey: pluginBindingSessionKey,
-      config: cfg,
-      fallbackAgentId: ctx.AgentId,
-    });
-    const blockedOwner = (expectedSessionId?: string): PluginBindingTranscriptOwner => ({
-      agentId: targetAgentId,
-      sessionKey: pluginBindingSessionKey,
-      ...(expectedSessionId ? { expectedSessionId } : {}),
-      transcriptWriteBlocked: true,
-    });
-    if (recorder.hasPersisted()) {
-      return blockedOwner();
-    }
-    let attemptedSessionId: string | undefined;
-    let lastOwner: PluginBindingTranscriptOwner | undefined;
-    for (let attempt = 0; attempt < 2; attempt += 1) {
-      const targetSessionStoreEntry = resolveSessionStoreLookup(
-        {
-          ...ctx,
-          CommandTargetSessionKey: undefined,
-          SessionKey: pluginBindingSessionKey,
-        },
-        cfg,
-      );
-      const targetSessionEntry = targetSessionStoreEntry.entry;
-      if (!targetSessionEntry || targetSessionEntry.sessionId === attemptedSessionId) {
-        break;
-      }
-      attemptedSessionId = targetSessionEntry.sessionId;
-      lastOwner = {
-        agentId: targetAgentId,
-        expectedSessionId: targetSessionEntry.sessionId,
-        sessionKey: pluginBindingSessionKey,
-      };
-      const result = await recorder.persistApproved({
-        target: {
-          sessionId: targetSessionEntry.sessionId,
-          sessionKey: pluginBindingSessionKey,
-          sessionEntry: targetSessionEntry,
-          ...(targetSessionStoreEntry.store ? { sessionStore: targetSessionStoreEntry.store } : {}),
-          storePath: targetSessionStoreEntry.storePath,
-          agentId: targetAgentId,
-          cwd: resolveAgentWorkspaceDir(cfg, targetAgentId),
-          config: cfg,
-        },
-        expectedSessionId: targetSessionEntry.sessionId,
-        retryIfUnpersisted: true,
-      });
-      if (result) {
-        return lastOwner;
-      }
-    }
-    if (!lastOwner) {
-      recorder.markBlocked();
-      return blockedOwner();
-    }
-    recorder.markBlocked();
-    logVerbose(`plugin-bound user-turn persistence skipped after the target session changed`);
-    return blockedOwner(lastOwner.expectedSessionId);
-  };
-
-  // Resolve automatic source-delivery suppression early so every outbound path
-  // below (plugin-binding notices, fast-abort, normal dispatch) honors it. The
-  // agent still processes inbound, but automatic replies/notices/indicators are
-  // blocked; explicit message tool sends remain available.
-  const sendPolicy = resolveSendPolicy({
-    cfg,
-    entry: sessionStoreEntry.entry,
-    sessionKey: sessionStoreEntry.sessionKey ?? sessionKey,
-    channel:
-      (shouldRouteToOriginating ? routeReplyChannel : undefined) ??
-      sessionStoreEntry.entry?.channel ??
-      replyRoute.channel ??
-      ctx.Surface ??
-      ctx.Provider ??
-      undefined,
-    chatType: sessionStoreEntry.entry?.chatType,
-  });
-  const {
-    globalPolicy,
-    globalProviderPolicy,
-    agentPolicy,
-    agentProviderPolicy,
-    profile,
-    providerProfile,
-    profileAlsoAllow,
-    providerProfileAlsoAllow,
-  } = resolveEffectiveToolPolicy({
-    config: cfg,
-    sessionKey: acpDispatchSessionKey,
-    agentId: sessionAgentId,
-  });
-  const chatType = normalizeChatType(ctx.ChatType);
-  const silentReplyConversationType = resolveRoutedPolicyConversationType(ctx);
-  const silentReplySurface = normalizeLowercaseStringOrEmpty(ctx.Surface ?? ctx.Provider);
-  const emptyFinalAllowedAsSilent =
-    silentReplyConversationType !== undefined &&
-    resolveSilentReplyPolicyFromPolicies({
-      conversationType: silentReplyConversationType,
-      defaultPolicy: cfg.agents?.defaults?.silentReply,
-      surfacePolicy: silentReplySurface
-        ? cfg.surfaces?.[silentReplySurface]?.silentReply
-        : undefined,
-    }) === "allow";
-  const configuredVisibleReplies =
-    chatType === "group" || chatType === "channel"
-      ? (cfg.messages?.groupChat?.visibleReplies ?? cfg.messages?.visibleReplies)
-      : cfg.messages?.visibleReplies;
-  const harnessDefaultVisibleReplies =
-    configuredVisibleReplies === undefined && chatType !== "group" && chatType !== "channel"
-      ? resolveHarnessSourceVisibleRepliesDefault({
-          cfg,
-          ctx,
-          entry: sessionStoreEntry.entry,
-          sessionAgentId,
-          sessionKey: acpDispatchSessionKey,
-          sessionStore: sessionStoreEntry.store,
-          turnModelOverride: resolveTurnModelOverride(params.replyOptions),
-        })
-      : undefined;
-  const effectiveVisibleReplies = configuredVisibleReplies ?? harnessDefaultVisibleReplies;
-  const prefersMessageToolDelivery =
-    params.replyOptions?.sourceReplyDeliveryMode === "message_tool_only" ||
-    (ctx.InboundEventKind === "room_event" && !isInternalWebchatTurn) ||
-    (params.replyOptions?.sourceReplyDeliveryMode === undefined &&
-      !isExplicitSourceReplyCommand(ctx, cfg) &&
-      (configuredVisibleReplies === "message_tool" ||
-        (!isInternalWebchatTurn && effectiveVisibleReplies === "message_tool")));
-  const runtimeProfileAlsoAllow = prefersMessageToolDelivery ? ["message"] : [];
-  const profilePolicy = mergeAlsoAllowPolicy(resolveToolProfilePolicy(profile), [
-    ...(profileAlsoAllow ?? []),
-    ...runtimeProfileAlsoAllow,
-  ]);
-  const providerProfilePolicy = mergeAlsoAllowPolicy(resolveToolProfilePolicy(providerProfile), [
-    ...(providerProfileAlsoAllow ?? []),
-    ...runtimeProfileAlsoAllow,
-  ]);
-  const groupResolution = resolveGroupSessionKey(ctx);
-  const messageProvider = resolveOriginMessageProvider({
-    originatingChannel: ctx.OriginatingChannel,
-    provider: ctx.Provider ?? ctx.Surface,
-  });
-  const groupPolicy = resolveGroupToolPolicy({
-    config: cfg,
-    sessionKey: acpDispatchSessionKey,
-    messageProvider,
-    groupId: groupResolution?.id,
-    groupChannel:
-      normalizeOptionalString(ctx.GroupChannel) ?? normalizeOptionalString(ctx.GroupSubject),
-    groupSpace: normalizeOptionalString(ctx.GroupSpace),
-    accountId: ctx.AccountId,
-    senderId: normalizeOptionalString(ctx.SenderId),
-    senderName: normalizeOptionalString(ctx.SenderName),
-    senderUsername: normalizeOptionalString(ctx.SenderUsername),
-    senderE164: normalizeOptionalString(ctx.SenderE164),
-  });
-  const subagentStore = resolveSubagentCapabilityStore(acpDispatchSessionKey, { cfg });
-  const subagentPolicy =
-    acpDispatchSessionKey &&
-    isSubagentEnvelopeSession(acpDispatchSessionKey, {
-      cfg,
-      store: subagentStore,
-    })
-      ? resolveSubagentToolPolicyForSession(cfg, acpDispatchSessionKey, {
-          store: subagentStore,
-        })
-      : undefined;
-  const inheritedToolPolicy = resolveInheritedToolPolicyForSession(cfg, acpDispatchSessionKey, {
-    store: subagentStore,
-  });
-  const messageToolAvailable = isToolAllowedByPolicies("message", [
-    profilePolicy,
-    providerProfilePolicy,
-    globalProviderPolicy,
-    agentProviderPolicy,
-    globalPolicy,
-    agentPolicy,
-    groupPolicy,
-    subagentPolicy,
-    inheritedToolPolicy,
-  ]);
-  const sourceReplyPolicy = resolveSourceReplyVisibilityPolicy({
-    cfg,
-    ctx,
-    requested: params.replyOptions?.sourceReplyDeliveryMode,
-    strictMessageToolOnly: ctx.InboundEventKind === "room_event" && !isInternalWebchatTurn,
-    sendPolicy,
-    suppressAcpChildUserDelivery,
-    explicitSuppressTyping: params.replyOptions?.suppressTyping === true,
-    shouldSuppressTyping,
-    messageToolAvailable,
-    defaultVisibleReplies: harnessDefaultVisibleReplies,
-  });
-  const {
-    sourceReplyDeliveryMode,
-    sessionStableSourceReplyDeliveryMode,
-    suppressAutomaticSourceDelivery,
-    suppressDelivery,
-    sendPolicyDenied,
-    deliverySuppressionReason,
-    suppressHookUserDelivery,
-    suppressHookReplyLifecycle,
-  } = sourceReplyPolicy;
-  const reasoningPayloadsEnabled = params.replyOptions?.reasoningPayloadsEnabled === true;
-  const commentaryPayloadsEnabled = params.replyOptions?.commentaryPayloadsEnabled === true;
-  const attachSourceReplyDeliveryMode = (
-    result: DispatchFromConfigResult,
-  ): DispatchFromConfigResult =>
-    sourceReplyDeliveryMode === "message_tool_only" || sendPolicyDenied
-      ? {
-          ...result,
-          ...(sourceReplyDeliveryMode === "message_tool_only" ? { sourceReplyDeliveryMode } : {}),
-          ...(sendPolicyDenied ? { sendPolicyDenied: true } : {}),
-        }
-      : result;
-  const explicitCommandTurnCtx = isExplicitSourceReplyCommand(ctx, cfg);
-  const unauthorizedTextSlashSourceReplyCtx =
-    (chatType === "group" || chatType === "channel") && isUnauthorizedTextSlashCommand(ctx);
-  const shouldDeliverPluginBindingReply =
-    !suppressAutomaticSourceDelivery ||
-    explicitCommandTurnCtx ||
-    (ctx.InboundEventKind !== "room_event" && !unauthorizedTextSlashSourceReplyCtx);
-
-  const durableSourceTurnId =
-    readChannelSourceTurnId(ctx) ??
-    (shouldMintChannelSourceTurnId(ctx.Provider ?? ctx.Surface)
-      ? buildChannelSourceTurnId({
-          provider: resolveOriginMessageProvider({
-            originatingChannel: replyRoute.channel,
-            provider: ctx.Provider ?? ctx.Surface,
-          }),
-          accountId: replyRoute.accountId,
-          conversationId: replyRoute.to,
-          messageId:
-            normalizeOptionalString(ctx.MessageSidFull) ?? normalizeOptionalString(ctx.MessageSid),
-        })
-      : undefined);
-  // Compute once before hooks. The prepared agent turn reuses this exact route-scoped id.
-  setChannelSourceTurnId(ctx, durableSourceTurnId);
-  if (isDuplicateRestartRecoverySource(sessionStoreEntry.entry, durableSourceTurnId)) {
-    // Process-local inbound dedupe cannot see provider redelivery after restart.
-    // Drop durable duplicates before any plugin dispatch hook can repeat effects.
-    recordProcessed("skipped", { reason: "duplicate" });
-    return attachSourceReplyDeliveryMode({
-      queuedFinal: false,
-      counts: dispatcher.getQueuedCounts(),
-    });
+  const gathered = await gatherDispatchRequest(params, messageAuditTerminal);
+  if (gathered.status === "complete") {
+    return gathered.result;
   }
 
-  const inboundDedupeClaim = claimInboundDedupe(ctx);
-  if (inboundDedupeClaim.status === "duplicate" || inboundDedupeClaim.status === "inflight") {
-    recordProcessed("skipped", { reason: "duplicate" });
-    return attachSourceReplyDeliveryMode({
-      queuedFinal: false,
-      counts: dispatcher.getQueuedCounts(),
-    });
-  }
-  const commitInboundDedupeIfClaimed = () => {
-    if (inboundDedupeClaim.status === "claimed") {
-      commitInboundDedupe(inboundDedupeClaim.key);
-    }
-  };
-  const releaseInboundDedupeIfClaimed = () => {
-    if (inboundDedupeClaim.status === "claimed") {
-      releaseInboundDedupe(inboundDedupeClaim.key);
-    }
-  };
-  const finishReplyOperationBusyDispatch = (opts?: {
-    dedupeDisposition?: "commit" | "release";
-    recordAgentDispatchCompleted?: boolean;
-    sessionMetadataChanges?: DispatchFromConfigResult["sessionMetadataChanges"];
-  }): DispatchFromConfigResult => {
-    void releasePreDispatchLifecycleAdmission(() => waitForReplyDispatcherIdle(dispatcher));
-    if (opts?.recordAgentDispatchCompleted) {
-      recordAgentDispatchCompleted("completed", { reason: "reply-operation-active" });
-    }
-    recordProcessed("skipped", { reason: "reply-operation-active" });
-    markIdle("message_completed");
-    if (opts?.dedupeDisposition === "release") {
-      releaseInboundDedupeIfClaimed();
-    } else {
-      commitInboundDedupeIfClaimed();
-    }
-    return attachSourceReplyDeliveryMode({
-      queuedFinal: false,
-      counts: dispatcher.getQueuedCounts(),
-      ...(opts?.sessionMetadataChanges
-        ? { sessionMetadataChanges: opts.sessionMetadataChanges }
-        : {}),
-    });
-  };
-  const finishReplyOperationAbortedDispatch = (): DispatchFromConfigResult => {
-    commitInboundDedupeIfClaimed();
-    recordProcessed("completed", { reason: "reply_operation_aborted" });
-    markIdle("message_completed");
-    completeDispatchReplyOperation();
-    return attachSourceReplyDeliveryMode({
-      queuedFinal: false,
-      counts: dispatcher.getQueuedCounts(),
-    });
-  };
+  const delivery = await prepareDispatchDelivery(gathered.state);
 
-  let pluginFallbackReason:
-    | "plugin-bound-fallback-missing-plugin"
-    | "plugin-bound-fallback-no-handler"
-    | undefined;
-  const emitMessageReceivedHooks = () => {
-    if (
-      ctx.SuppressMessageReceivedHooks !== true &&
-      hookRunner?.hasHooks("message_received") === true
-    ) {
-      const messageReceivedHookContext = buildMessageReceivedHookContext();
-      fireAndForgetHook(
-        hookRunner.runMessageReceived(
-          toPluginMessageReceivedEvent(messageReceivedHookContext),
-          toPluginMessageContext(messageReceivedHookContext),
-        ),
-        "dispatch-from-config: message_received plugin hook failed",
-      );
-    }
-    if (ctx.SuppressMessageReceivedHooks !== true && sessionKey) {
-      const messageReceivedHookContext = buildMessageReceivedHookContext();
-      fireAndForgetHook(
-        triggerInternalHook(
-          createInternalHookEvent("message", "received", sessionKey, {
-            ...toInternalMessageReceivedContext(messageReceivedHookContext),
-            timestamp,
-          }),
-        ),
-        "dispatch-from-config: message_received internal hook failed",
-      );
-    }
-  };
-  markProcessing();
-  if (await capturePendingConversationTurnReply({ cfg, ctx })) {
-    emitMessageReceivedHooks();
-    commitInboundDedupeIfClaimed();
-    recordProcessed("completed", { reason: "conversation-turn-reply" });
-    markIdle("message_completed");
-    return attachSourceReplyDeliveryMode({
-      queuedFinal: false,
-      counts: dispatcher.getQueuedCounts(),
-      observedReplyDelivery: true,
-    });
+  const context = await prepareDispatchOperationContext(delivery.state);
+  if (context.status === "complete") {
+    return context.result;
   }
+
+  const errorState = context.state;
   try {
-    const abortRuntime = params.fastAbortResolver ? null : await loadAbortRuntime();
-    const fastAbortResolver = params.fastAbortResolver ?? abortRuntime?.tryFastAbortFromMessage;
-    const formatAbortReplyTextResolver =
-      params.formatAbortReplyTextResolver ?? abortRuntime?.formatAbortReplyText;
-    if (!fastAbortResolver || !formatAbortReplyTextResolver) {
-      throw new Error("abort runtime unavailable");
-    }
-    const fastAbort = await fastAbortResolver({ ctx, cfg });
-    if (fastAbort.handled) {
-      if (pluginOwnedBinding) {
-        touchConversationBindingRecord(pluginOwnedBinding.bindingId);
-      }
-      emitMessageReceivedHooks();
-      let queuedFinal = false;
-      let routedFinalCount = 0;
-      if (!suppressDelivery) {
-        const selectedModel = resolveSessionModelRef(cfg, sessionStoreEntry.entry, sessionAgentId);
-        const modelSelection = {
-          ...selectedModel,
-          thinkLevel: sessionStoreEntry.entry?.thinkingLevel,
-        };
-        const responsePrefixContext = {
-          identityName: normalizeOptionalString(resolveAgentIdentity(cfg, sessionAgentId)?.name),
-          provider: selectedModel.provider,
-          model: extractShortModelName(selectedModel.model),
-          modelFull: `${selectedModel.provider}/${selectedModel.model}`,
-          thinkingLevel: modelSelection.thinkLevel ?? "off",
-        };
-        const payload = {
-          text: formatAbortReplyTextResolver(fastAbort.stoppedSubagents, fastAbort.rejectionReason),
-        } satisfies ReplyPayload;
-        // Routed delivery owns its destination-scoped prefix. Direct dispatchers already own
-        // their prefix, so seed that live context only when no cross-channel route is used.
-        const result = await routeReplyToOriginating(payload, { responsePrefixContext });
-        if (result) {
-          queuedFinal = result.ok;
-          if (isRoutedReplyDelivered(result)) {
-            routedFinalCount += 1;
-          }
-          if (!result.ok) {
-            logVerbose(
-              `dispatch-from-config: route-reply (abort) failed: ${result.error ?? "unknown error"}`,
-            );
-          }
-        } else {
-          markInboundDedupeReplayUnsafe();
-          params.replyOptions?.onModelSelected?.(modelSelection);
-          queuedFinal = dispatcher.sendFinalReply(payload);
-        }
-      } else {
-        logVerbose(
-          `dispatch-from-config: fast_abort reply suppressed by ${deliverySuppressionReason} (session=${sessionKey ?? "unknown"})`,
-        );
-      }
-      const counts = dispatcher.getQueuedCounts();
-      counts.final += routedFinalCount;
-      recordProcessed("completed", { reason: "fast_abort" });
-      markIdle("message_completed");
-      commitInboundDedupeIfClaimed();
-      completeDispatchReplyOperation();
-      return attachSourceReplyDeliveryMode({ queuedFinal, counts });
-    }
-    // Own the session before plugin-bound handlers or message hooks can perform
-    // work. Fast abort and inbound dedupe intentionally remain ahead of this gate.
-    const preDispatchAcquisition = await ensureDispatchReplyOperation("pre_dispatch");
-    if (preDispatchAcquisition.status === "aborted") {
-      return finishReplyOperationAbortedDispatch();
-    }
-    if (preDispatchAcquisition.status === "busy") {
-      return finishReplyOperationBusyDispatch({ dedupeDisposition: "release" });
+    const operation = await prepareDispatchOperation(context.state);
+    if (operation.status === "complete") {
+      return operation.result;
     }
 
-    if (pluginOwnedBinding) {
-      if (isPreDispatchOperationAborted()) {
-        return finishReplyOperationAbortedDispatch();
-      }
-      touchConversationBindingRecord(pluginOwnedBinding.bindingId);
-      if (shouldBypassPluginOwnedBindingForCommand(ctx, cfg)) {
-        logVerbose(
-          `plugin-bound inbound command escaped plugin binding (plugin=${pluginOwnedBinding.pluginId} session=${sessionKey ?? "unknown"}); falling through to command processing`,
-        );
-      } else if (sendPolicyDenied || (suppressDelivery && !suppressAutomaticSourceDelivery)) {
-        // Plugin-bound inbound handlers typically emit outbound replies we
-        // cannot rewind. When automatic delivery is explicitly denied, skip the
-        // plugin claim and fall through to normal suppressed agent processing.
-        // message_tool_only is the normal visible-reply mode for group chats and
-        // must still let the bound plugin own the turn unless sendPolicy denied it.
-        logVerbose(
-          `plugin-bound inbound skipped under ${deliverySuppressionReason} (plugin=${pluginOwnedBinding.pluginId} session=${sessionKey ?? "unknown"}); falling through to suppressed agent processing`,
-        );
-      } else {
-        logVerbose(
-          `plugin-bound inbound routed to ${pluginOwnedBinding.pluginId} conversation=${pluginOwnedBinding.conversationId}`,
-        );
-        // Bound native runtimes need the current owner decision, not stale bind-time identity.
-        // The resolver folds internal operator.admin authority into this owner decision.
-        const bindingAuthorization = resolveCommandAuthorization({
-          ctx,
-          cfg,
-          commandAuthorized: ctx.CommandAuthorized,
-        });
-        const targetedClaimOutcome = hookRunner?.runInboundClaimForPluginOutcome
-          ? await (async () => {
-              await prepareHookMediaMetadata();
-              if (isPreDispatchOperationAborted()) {
-                throw new DispatchReplyOperationAbortedError();
-              }
-              const authorizedInboundClaimEvent = {
-                ...inboundClaimEvent,
-                senderIsOwner: bindingAuthorization.senderIsOwner,
-              };
-              return await runWithDispatchLifecycleAdmission(
-                async () =>
-                  await hookRunner.runInboundClaimForPluginOutcome(
-                    pluginOwnedBinding.pluginId,
-                    authorizedInboundClaimEvent,
-                    { ...inboundClaimContext, pluginBinding: pluginOwnedBinding },
-                  ),
-              );
-            })()
-          : (() => {
-              const pluginLoaded =
-                getGlobalPluginRegistry()?.plugins.some(
-                  (plugin) =>
-                    plugin.id === pluginOwnedBinding.pluginId && plugin.status === "loaded",
-                ) ?? false;
-              return pluginLoaded
-                ? ({ status: "no_handler" } as const)
-                : ({ status: "missing_plugin" } as const);
-            })();
-        if (isPreDispatchOperationAborted()) {
-          return finishReplyOperationAbortedDispatch();
-        }
-
-        switch (targetedClaimOutcome.status) {
-          case "handled": {
-            const transcriptOwner = await persistPluginBindingUserTurn();
-            if (targetedClaimOutcome.result.reply && shouldDeliverPluginBindingReply) {
-              // A bound plugin's reply is the explicit output for this claimed turn,
-              // not an automatic agent final; message-tool-only suppression must not
-              // turn normal user-request bindings into silent channel responses.
-              // Ambient room events keep the same privacy guard as final replies.
-              await deliverBindingPayload(
-                targetedClaimOutcome.result.reply,
-                "terminal",
-                transcriptOwner,
-              );
-            }
-            markIdle("plugin_binding_dispatch");
-            recordProcessed("completed", { reason: "plugin-bound-handled" });
-            commitInboundDedupeIfClaimed();
-            completeDispatchReplyOperation();
-            return attachSourceReplyDeliveryMode({
-              queuedFinal: false,
-              counts: dispatcher.getQueuedCounts(),
-            });
-          }
-          case "missing_plugin":
-          case "no_handler": {
-            pluginFallbackReason =
-              targetedClaimOutcome.status === "missing_plugin"
-                ? "plugin-bound-fallback-missing-plugin"
-                : "plugin-bound-fallback-no-handler";
-            const isUnmentionedGroupFallback =
-              (chatType === "group" || chatType === "channel") &&
-              ctx.WasMentioned === false &&
-              !explicitCommandTurnCtx;
-            const shouldSuppressUnmentionedFallback =
-              isUnmentionedGroupFallback && ctx.GroupRequireMention !== false;
-            if (shouldSuppressUnmentionedFallback) {
-              markIdle("plugin_binding_fallback_unmentioned");
-              recordProcessed("completed", { reason: pluginFallbackReason });
-              commitInboundDedupeIfClaimed();
-              completeDispatchReplyOperation();
-              return attachSourceReplyDeliveryMode({
-                queuedFinal: false,
-                counts: dispatcher.getQueuedCounts(),
-              });
-            }
-            if (!hasShownPluginBindingFallbackNotice(pluginOwnedBinding.bindingId)) {
-              const didSendNotice = await sendBindingNotice(
-                { text: buildPluginBindingUnavailableText(pluginOwnedBinding) },
-                "additive",
-              );
-              if (didSendNotice) {
-                markPluginBindingFallbackNoticeShown(pluginOwnedBinding.bindingId);
-              }
-            }
-            break;
-          }
-          case "declined": {
-            const transcriptOwner = await persistPluginBindingUserTurn();
-            await sendBindingNotice(
-              { text: buildPluginBindingDeclinedText(pluginOwnedBinding) },
-              "terminal",
-              transcriptOwner,
-            );
-            markIdle("plugin_binding_declined");
-            recordProcessed("completed", { reason: "plugin-bound-declined" });
-            commitInboundDedupeIfClaimed();
-            completeDispatchReplyOperation();
-            return attachSourceReplyDeliveryMode({
-              queuedFinal: false,
-              counts: dispatcher.getQueuedCounts(),
-            });
-          }
-          case "error": {
-            const transcriptOwner = await persistPluginBindingUserTurn();
-            logVerbose(
-              `plugin-bound inbound claim failed for ${pluginOwnedBinding.pluginId}: ${targetedClaimOutcome.error}`,
-            );
-            await sendBindingNotice(
-              { text: buildPluginBindingErrorText(pluginOwnedBinding) },
-              "terminal",
-              transcriptOwner,
-            );
-            markIdle("plugin_binding_error");
-            recordProcessed("completed", { reason: "plugin-bound-error" });
-            commitInboundDedupeIfClaimed();
-            completeDispatchReplyOperation();
-            return attachSourceReplyDeliveryMode({
-              queuedFinal: false,
-              counts: dispatcher.getQueuedCounts(),
-            });
-          }
-        }
-      }
+    const route = await chooseDispatchRoute(operation.state);
+    if (route.status === "complete") {
+      return route.result;
     }
 
-    emitMessageReceivedHooks();
+    const execution = await prepareDispatchExecution(route.state);
 
-    const shouldSuppressDefaultToolProgressMessages = () => !shouldEmitVerboseProgress();
-    const shouldSendVerboseProgressMessages = () => !shouldSuppressDefaultToolProgressMessages();
-    const shouldSendToolSummaries = () => shouldSendVerboseProgressMessages();
-    const shouldSendToolStartStatuses = false;
-    const notifiedSessionMetadataChangeKeys = new Set<string>();
-    let sessionMetadataChangesForResult: CommandSessionMetadataChange[] | undefined;
-    const notifySessionMetadataChanges = (
-      changes: CommandSessionMetadataChange[] | undefined,
-    ): void => {
-      if (!changes?.length) {
-        return;
-      }
-      const freshChanges: CommandSessionMetadataChange[] = [];
-      for (const change of changes) {
-        const key = JSON.stringify([change.sessionKey, change.agentId ?? null, change.reason]);
-        if (notifiedSessionMetadataChangeKeys.has(key)) {
-          continue;
-        }
-        notifiedSessionMetadataChangeKeys.add(key);
-        freshChanges.push(change);
-      }
-      if (freshChanges.length === 0) {
-        return;
-      }
-      sessionMetadataChangesForResult = [
-        ...(sessionMetadataChangesForResult ?? []),
-        ...freshChanges,
-      ];
-      params.onSessionMetadataChanges?.(freshChanges);
-    };
-    const shouldDeliverVerboseProgressDespiteSourceSuppression = () =>
-      suppressAutomaticSourceDelivery &&
-      sourceReplyDeliveryMode === "message_tool_only" &&
-      ctx.InboundEventKind !== "room_event" &&
-      !sendPolicyDenied &&
-      shouldEmitVerboseProgress() &&
-      shouldSendVerboseProgressMessages();
-    const shouldDeliverForcedToolProgressDespiteSourceSuppression = () =>
-      suppressAutomaticSourceDelivery &&
-      sourceReplyDeliveryMode === "message_tool_only" &&
-      ctx.InboundEventKind !== "room_event" &&
-      !sendPolicyDenied &&
-      params.replyOptions?.forceToolResultProgress === true;
-    const shouldDeliverFastModeAutoProgressDespiteSourceSuppression = () =>
-      suppressAutomaticSourceDelivery &&
-      sourceReplyDeliveryMode === "message_tool_only" &&
-      ctx.InboundEventKind !== "room_event" &&
-      !sendPolicyDenied;
-    let finalReplyDeliveryStarted = false;
-    const hasExecApprovalPayload = (payload: ReplyPayload) => {
-      const execApproval =
-        payload.channelData &&
-        typeof payload.channelData === "object" &&
-        !Array.isArray(payload.channelData)
-          ? payload.channelData.execApproval
-          : undefined;
-      return execApproval && typeof execApproval === "object" && !Array.isArray(execApproval);
-    };
-    const hasAskUserPayload = (payload: ReplyPayload) => {
-      const askUser = payload.channelData?.askUser;
-      return askUser && typeof askUser === "object" && !Array.isArray(askUser);
-    };
-    const readAskUserQuestionId = (payload: ReplyPayload) => {
-      const askUser = payload.channelData?.askUser;
-      if (!askUser || typeof askUser !== "object" || Array.isArray(askUser)) {
-        return undefined;
-      }
-      const questionId = (askUser as { questionId?: unknown }).questionId;
-      return typeof questionId === "string" ? questionId : undefined;
-    };
-    const shouldSuppressLateTextOnlyToolProgress = (payload: ReplyPayload) => {
-      if (!finalReplyDeliveryStarted) {
-        return false;
-      }
-      const reply = resolveSendableOutboundReplyParts(payload);
-      return !reply.hasMedia && !hasExecApprovalPayload(payload) && !hasAskUserPayload(payload);
-    };
-    // Durable inter-tool commentary lane: with verbose progress on, preamble
-    // items become standalone progress messages like tool summaries. The latest
-    // text per item id is buffered (snapshot producers re-emit the same item)
-    // and flushed when the producer moves on, always before the final reply.
-    let pendingCommentaryProgress: { itemId?: string; text: string } | null = null;
-    const deliverCommentaryProgressMessage = async (text: string) => {
-      if (!shouldSendToolSummaries() || shouldSuppressProgressDelivery()) {
-        return;
-      }
-      const payload: ReplyPayload = { text: `💬 ${text}` };
-      if (shouldSuppressLateTextOnlyToolProgress(payload)) {
-        return;
-      }
-      if (shouldRouteToOriginating) {
-        await sendPayloadAsync(payload, undefined, false);
-      } else {
-        markInboundDedupeReplayUnsafe();
-        dispatcher.sendToolResult(payload);
-      }
-    };
-    const flushPendingCommentaryProgress = async () => {
-      const pending = pendingCommentaryProgress;
-      pendingCommentaryProgress = null;
-      const text = pending?.text.trim();
-      if (!text) {
-        return;
-      }
-      await deliverCommentaryProgressMessage(text);
-    };
-    const noteCommentaryProgress = async (payload: { itemId?: string; progressText?: string }) => {
-      const itemId = payload.itemId?.trim() || undefined;
-      const text = payload.progressText ?? "";
-      const updatesBufferedItem =
-        pendingCommentaryProgress !== null &&
-        pendingCommentaryProgress.itemId !== undefined &&
-        pendingCommentaryProgress.itemId === itemId;
-      if (!text.trim()) {
-        // Empty commentary with an item id means the producer retracted that
-        // item; drop it if it has not been sent yet.
-        if (updatesBufferedItem) {
-          pendingCommentaryProgress = null;
-        }
-        return;
-      }
-      if (pendingCommentaryProgress && !updatesBufferedItem) {
-        await flushPendingCommentaryProgress();
-      }
-      pendingCommentaryProgress = { itemId, text };
-    };
-    const shouldSuppressMessageToolOnlyTextErrorProgress = (payload: ReplyPayload) => {
-      if (
-        sourceReplyDeliveryMode !== "message_tool_only" ||
-        shouldEmitFullVerboseProgress() ||
-        payload.isError !== true
-      ) {
-        return false;
-      }
-      const reply = resolveSendableOutboundReplyParts(payload);
-      return !reply.hasMedia && !hasExecApprovalPayload(payload);
-    };
-    const sendFinalPayload = async (
-      payload: ReplyPayload,
-      options: { abortSignal?: AbortSignal; deliveryId?: string } = {},
-    ): Promise<{
-      queuedFinal: boolean;
-      routedFinalCount: number;
-      dispatcherOutcome?: Promise<ReplyDispatchDeliveryOutcome>;
-    }> => {
-      const abortSignal = options.abortSignal ?? getDispatchAbortSignal();
-      const throwIfFinalDeliveryAborted = () => {
-        if (abortSignal?.aborted) {
-          throw new DispatchReplyOperationAbortedError();
-        }
-      };
-      throwIfFinalDeliveryAborted();
-      // Trailing commentary must land ahead of the final answer.
-      await flushPendingCommentaryProgress();
-      throwIfFinalDeliveryAborted();
-      const payloadMetadata = getReplyPayloadMetadata(payload);
-      const sourceReplySessionBinding = resolvePreparedTranscriptBinding(
-        payloadMetadata?.sourceReplyTranscriptMirror?.sessionKey,
-      );
-      const sourceReplyTranscriptMirror = payloadMetadata?.sourceReplyTranscriptMirror
-        ? {
-            ...payloadMetadata.sourceReplyTranscriptMirror,
-            ...(sourceReplySessionBinding
-              ? { expectedSessionId: sourceReplySessionBinding.sessionId }
-              : {}),
-            storePath: sourceReplySessionBinding?.storePath ?? sessionStoreEntry.storePath,
-          }
-        : undefined;
-      const hasTranscriptOwner =
-        payloadMetadata?.assistantMessageIndex !== undefined ||
-        payloadMetadata?.assistantTranscriptOwned === true;
-      const hasVisibleFinalContent = hasOutboundReplyContent(payload, { trimText: true });
-      if (hasVisibleFinalContent) {
-        markInboundDedupeReplayUnsafe();
-        finalReplyDeliveryStarted = true;
-      }
-      const ttsPayload =
-        payload.isReasoning === true || payload.isCommentary === true
-          ? payload
-          : await maybeApplyTtsWithFinalizationLease({
-              payload,
-              cfg,
-              channel: deliveryChannel,
-              kind: "final",
-              ttsAuto: sessionTtsAuto,
-              agentId: sessionAgentId,
-              accountId: replyRoute.accountId,
-            });
-      throwIfFinalDeliveryAborted();
-      const normalizedPayload = await normalizeReplyMediaPayload(ttsPayload);
-      throwIfFinalDeliveryAborted();
-      const result = await routeReplyToOriginating(normalizedPayload, {
-        abortSignal,
-        kind: "final",
-        ...(hasTranscriptOwner ? { mirror: false } : {}),
-      });
-      if (result) {
-        if (!result.ok) {
-          logVerbose(
-            `dispatch-from-config: route-reply (final) failed: ${result.error ?? "unknown error"}`,
-          );
-        }
-        if (isRoutedReplyDelivered(result)) {
-          await mirrorDeliveredReplyToTranscript({
-            metadata: sourceReplyTranscriptMirror,
-            cfg,
-          });
-        }
-        return {
-          queuedFinal: result.ok,
-          routedFinalCount: isRoutedReplyDelivered(result) ? 1 : 0,
-        };
-      }
-      throwIfFinalDeliveryAborted();
-      const transcriptMirrorSessionKey =
-        acpDispatchSessionKey ?? sessionStoreEntry.sessionKey ?? sessionKey;
-      const transcriptMirrorSourceId =
-        normalizeOptionalString(messageIdForHook) ??
-        normalizeOptionalString(params.replyOptions?.runId);
-      const transcriptMirrorSessionBinding = resolvePreparedTranscriptBinding(
-        transcriptMirrorSessionKey,
-      );
-      const transcriptMirror =
-        sourceReplyTranscriptMirror ??
-        (normalizedCurrentSurface === "slack" &&
-        hasVisibleFinalContent &&
-        transcriptMirrorSessionKey
-          ? transcriptMirrorForDeliveredPayload(
-              {
-                sessionKey: transcriptMirrorSessionKey,
-                agentId: sessionAgentId,
-                ...(transcriptMirrorSessionBinding
-                  ? { expectedSessionId: transcriptMirrorSessionBinding.sessionId }
-                  : {}),
-                storePath: transcriptMirrorSessionBinding?.storePath ?? sessionStoreEntry.storePath,
-                preferText: true,
-                ...(hasTranscriptOwner ? { transcriptOwner: true } : {}),
-                idempotencyKey: transcriptMirrorSourceId
-                  ? `channel-final:${transcriptMirrorSourceId}:${options.deliveryId ?? "single"}`
-                  : undefined,
-                deliveryMirror: {
-                  kind: "channel-final",
-                  ...(transcriptMirrorSourceId
-                    ? { sourceMessageId: transcriptMirrorSourceId }
-                    : {}),
-                },
-              },
-              normalizedPayload,
-            )
-          : undefined);
-      markInboundDedupeReplayUnsafe();
-      const finalOutcomeBefore = transcriptMirror
-        ? getDispatcherFinalOutcomeCounts(dispatcher)
-        : undefined;
-      const finalDeliveryCapture = transcriptMirror ? {} : undefined;
-      const deliveredTranscriptMirror = transcriptMirror
-        ? captureDeliveredTranscriptMirror({
-            dispatcher,
-            metadata: transcriptMirror,
-            deliveryId: options.deliveryId,
-            captureToken: finalDeliveryCapture,
-          })
-        : undefined;
-      if (finalDeliveryCapture) {
-        setReplyPayloadMetadata(normalizedPayload, { finalDeliveryCapture });
-      }
-      const deliveryOutcome = captureReplyDispatchDeliveryOutcome(normalizedPayload);
-      const queuedFinal = dispatcher.sendFinalReply(normalizedPayload);
-      const dispatcherOutcome =
-        queuedFinal && deliveryOutcome.isTracked() ? deliveryOutcome.promise : undefined;
-      if (queuedFinal && deliveredTranscriptMirror && finalOutcomeBefore) {
-        // The common settle owner runs this after successful delivery or
-        // cancellation. Keeping reconciliation out of the reply operation lets a
-        // newer foreground turn settle without creating an operation/idle cycle.
-        registerReplyDispatcherSettledTask(dispatcher, () =>
-          mirrorTranscriptAfterDispatcherSettled({
-            dispatcher,
-            before: finalOutcomeBefore,
-            metadata: deliveredTranscriptMirror,
-            cfg,
-          }),
-        );
-      }
-      return {
-        queuedFinal,
-        routedFinalCount: 0,
-        ...(queuedFinal && dispatcherOutcome ? { dispatcherOutcome } : {}),
-      };
-    };
-
-    // Run before_dispatch hook — let plugins inspect or handle before model dispatch.
-    if (hookRunner?.hasHooks("before_dispatch")) {
-      const beforeDispatchResult = await traceReplyPhase("reply.before_dispatch_hooks", () =>
-        runWithDispatchLifecycleAdmission(
-          async () =>
-            await runWithDispatchAbortSignal(
-              getPreDispatchAbortSignal(),
-              () =>
-                hookRunner.runBeforeDispatch(
-                  {
-                    content: hookContext.content,
-                    body: hookContext.bodyForAgent ?? hookContext.body,
-                    channel: hookContext.channelId,
-                    sessionKey: sessionStoreEntry.sessionKey ?? sessionKey,
-                    senderId: hookContext.senderId,
-                    replyToId: hookContext.replyToId,
-                    replyToIdFull: hookContext.replyToIdFull,
-                    replyToBody: hookContext.replyToBody,
-                    replyToSender: hookContext.replyToSender,
-                    replyToIsQuote: hookContext.replyToIsQuote,
-                    isGroup: hookContext.isGroup,
-                    timestamp: hookContext.timestamp,
-                  },
-                  {
-                    channelId: hookContext.channelId,
-                    accountId: hookContext.accountId,
-                    conversationId: inboundClaimContext.conversationId,
-                    sessionKey: sessionStoreEntry.sessionKey ?? sessionKey,
-                    senderId: hookContext.senderId,
-                    replyToId: hookContext.replyToId,
-                    replyToIdFull: hookContext.replyToIdFull,
-                    replyToBody: hookContext.replyToBody,
-                    replyToSender: hookContext.replyToSender,
-                    replyToIsQuote: hookContext.replyToIsQuote,
-                  },
-                ),
-              trackDispatchLifecycleWork,
-            ),
-        ),
-      );
-      if (beforeDispatchResult?.handled) {
-        const text = beforeDispatchResult.text;
-        let queuedFinal = false;
-        let routedFinalCount = 0;
-        if (text && !suppressDelivery) {
-          const handledReply = await sendFinalPayload(
-            { text },
-            {
-              abortSignal: getPreDispatchAbortSignal(),
-              deliveryId: "before-dispatch",
-            },
-          );
-          queuedFinal = handledReply.queuedFinal;
-          routedFinalCount += handledReply.routedFinalCount;
-        }
-        const counts = dispatcher.getQueuedCounts();
-        counts.final += routedFinalCount;
-        recordProcessed("completed", { reason: "before_dispatch_handled" });
-        markIdle("message_completed");
-        commitInboundDedupeIfClaimed();
-        completeDispatchReplyOperation();
-        return attachSourceReplyDeliveryMode({ queuedFinal, counts });
-      }
+    const executed = await executeDispatch(execution.state);
+    if (executed.status === "complete") {
+      return executed.result;
     }
 
-    if (hookRunner?.hasHooks("reply_dispatch")) {
-      const replyDispatchResult = await traceReplyPhase("reply.reply_dispatch_hooks", () =>
-        runWithDispatchLifecycleAdmission(
-          async () =>
-            await runWithDispatchAbortSignal(
-              getPreDispatchAbortSignal(),
-              () =>
-                hookRunner.runReplyDispatch(
-                  createReplyDispatchEvent({
-                    ctx,
-                    runId: params.replyOptions?.runId,
-                    sessionKey: acpDispatchSessionKey,
-                    toolsAllow: params.replyOptions?.toolsAllow,
-                    images: params.replyOptions?.images,
-                    inboundAudio,
-                    sessionTtsAuto,
-                    ttsChannel: deliveryChannel,
-                    suppressUserDelivery: suppressHookUserDelivery,
-                    suppressReplyLifecycle: suppressHookReplyLifecycle,
-                    sourceReplyDeliveryMode,
-                    shouldRouteToOriginating,
-                    originatingChannel: routeReplyChannel,
-                    originatingTo: routeReplyTo,
-                    originatingAccountId: replyContextAccountId,
-                    originatingThreadId: routeReplyThreadId,
-                    originatingChatType: replyRoute.chatType,
-                    shouldSendToolSummaries,
-                    sendPolicy,
-                  }),
-                  {
-                    cfg,
-                    dispatcher: dispatchHookDispatcher,
-                    abortSignal: getPreDispatchAbortSignal() ?? params.replyOptions?.abortSignal,
-                    onReplyStart: params.replyOptions?.onReplyStart,
-                    recordProcessed,
-                    markIdle,
-                  },
-                ),
-              trackDispatchLifecycleWork,
-            ),
-        ),
-      );
-      if (replyDispatchResult?.handled) {
-        commitInboundDedupeIfClaimed();
-        completeDispatchReplyOperation();
-        return attachSourceReplyDeliveryMode({
-          queuedFinal: replyDispatchResult.queuedFinal,
-          counts: replyDispatchResult.counts,
-        });
-      }
-    }
-
-    const dispatchAcquisition = await ensureDispatchReplyOperation("dispatch");
-    if (dispatchAcquisition.status === "aborted") {
-      return finishReplyOperationAbortedDispatch();
-    }
-    if (dispatchAcquisition.status === "busy") {
-      return finishReplyOperationBusyDispatch({ dedupeDisposition: "release" });
-    }
-
-    // When automatic source delivery is suppressed, still let the agent process
-    // the inbound message (context, memory, tool calls) but suppress automatic
-    // outbound source delivery.
-    if (suppressDelivery) {
-      logVerbose(
-        `Delivery suppressed by ${deliverySuppressionReason} for session ${sessionStoreEntry.sessionKey ?? sessionKey ?? "unknown"} — agent will still process the message`,
-      );
-    }
-
-    const toolStartStatusesSent = new Set<string>();
-    let toolStartStatusCount = 0;
-    let didSendPlanStatusNotice = false;
-    const normalizeWorkingLabel = (label: string) => {
-      const collapsed = label.replace(/\s+/g, " ").trim();
-      if (collapsed.length <= 80) {
-        return collapsed;
-      }
-      return `${truncateUtf16Safe(collapsed, 77).trimEnd()}...`;
-    };
-    const formatPlanUpdateText = (payload: { explanation?: string; steps?: AgentPlanStep[] }) => {
-      const explanation = payload.explanation?.replace(/\s+/g, " ").trim();
-      const steps = (payload.steps ?? [])
-        .map((entry) => ({ step: entry.step.replace(/\s+/g, " ").trim(), status: entry.status }))
-        .filter((entry) => entry.step);
-      if (steps.length > 0) {
-        return formatPlanChecklistLines(steps, {
-          maxLines: steps.length,
-          maxLineChars: 120,
-        }).join("\n");
-      }
-      return explanation || "Planning next steps.";
-    };
-    const maybeSendWorkingStatus = async (label: string): Promise<void> => {
-      if (shouldSuppressProgressDelivery()) {
-        return;
-      }
-      const normalizedLabel = normalizeWorkingLabel(label);
-      if (
-        !shouldEmitVerboseProgress() ||
-        !shouldSendToolStartStatuses ||
-        !normalizedLabel ||
-        toolStartStatusCount >= 2 ||
-        toolStartStatusesSent.has(normalizedLabel)
-      ) {
-        return;
-      }
-      toolStartStatusesSent.add(normalizedLabel);
-      toolStartStatusCount += 1;
-      const payload: ReplyPayload = {
-        text: `Working: ${normalizedLabel}`,
-      };
-      if (shouldRouteToOriginating) {
-        await sendPayloadAsync(payload, undefined, false);
-        return;
-      }
-      markInboundDedupeReplayUnsafe();
-      dispatcher.sendToolResult(payload);
-    };
-    const sendPlanUpdate = async (payload: {
-      explanation?: string;
-      steps?: AgentPlanStep[];
-    }): Promise<void> => {
-      if (
-        shouldSuppressProgressDelivery() ||
-        !shouldSendVerboseProgressMessages() ||
-        didSendPlanStatusNotice
-      ) {
-        return;
-      }
-      didSendPlanStatusNotice = true;
-      const replyPayload: ReplyPayload = {
-        text: formatPlanUpdateText(payload),
-        isStatusNotice: true,
-      };
-      if (shouldRouteToOriginating) {
-        await sendPayloadAsync(replyPayload, undefined, false);
-        return;
-      }
-      markInboundDedupeReplayUnsafe();
-      dispatcher.sendToolResult(replyPayload);
-    };
-    const summarizeApprovalLabel = (payload: {
-      status?: string;
-      command?: string;
-      message?: string;
-    }) => {
-      if (payload.status === "pending") {
-        const command = normalizeOptionalString(payload.command);
-        if (command) {
-          return normalizeWorkingLabel(`awaiting approval: ${command}`);
-        }
-        return "awaiting approval";
-      }
-      if (payload.status === "unavailable") {
-        const message = normalizeOptionalString(payload.message);
-        if (message) {
-          return normalizeWorkingLabel(message);
-        }
-        return "approval unavailable";
-      }
-      return "";
-    };
-    const summarizePatchLabel = (payload: { summary?: string; title?: string }) => {
-      const summary = normalizeOptionalString(payload.summary);
-      if (summary) {
-        return normalizeWorkingLabel(summary);
-      }
-      const title = normalizeOptionalString(payload.title);
-      if (title) {
-        return normalizeWorkingLabel(title);
-      }
-      return "";
-    };
-    // Track accumulated block text for TTS generation after streaming completes.
-    // When block streaming succeeds, there's no final reply, so we need to generate
-    // TTS audio separately from the accumulated block content.
-    let accumulatedBlockText = "";
-    let accumulatedBlockTtsText = "";
-    let blockCount = 0;
-    const cleanBlockTtsDirectiveText = shouldCleanTtsDirectiveText({
-      cfg,
-      ttsAuto: sessionTtsAuto,
-      agentId: sessionAgentId,
-      channelId: deliveryChannel,
-      accountId: replyRoute.accountId,
-    })
-      ? createTtsDirectiveTextStreamCleaner()
-      : undefined;
-
-    const resolveToolDeliveryPayload = (payload: ReplyPayload): ReplyPayload | null => {
-      if (
-        shouldSuppressLocalExecApprovalPrompt({
-          channel: normalizeMessageChannel(ctx.Surface ?? ctx.Provider),
-          cfg,
-          accountId: ctx.AccountId,
-          payload,
-        })
-      ) {
-        return null;
-      }
-      if (shouldSendToolSummaries()) {
-        return payload;
-      }
-      const execApproval =
-        payload.channelData &&
-        typeof payload.channelData === "object" &&
-        !Array.isArray(payload.channelData)
-          ? payload.channelData.execApproval
-          : undefined;
-      if (execApproval && typeof execApproval === "object" && !Array.isArray(execApproval)) {
-        return payload;
-      }
-      if (hasAskUserPayload(payload)) {
-        return payload;
-      }
-      if (isFastModeAutoProgressPayload(payload)) {
-        return payload;
-      }
-      // Group/native flows intentionally suppress tool summary text, but media-only
-      // tool results (for example TTS audio) must still be delivered.
-      const hasMedia = resolveSendableOutboundReplyParts(payload).hasMedia;
-      if (!hasMedia) {
-        return null;
-      }
-      return { ...payload, text: undefined };
-    };
-    const typing = resolveRunTypingPolicy({
-      requestedPolicy: params.replyOptions?.typingPolicy,
-      suppressTyping: sourceReplyPolicy.suppressTyping,
-      originatingChannel: routeReplyChannel,
-      systemEvent: shouldRouteToOriginating,
-    });
-    const shouldSuppressProgressDelivery = () =>
-      sendPolicyDenied ||
-      (suppressDelivery && !shouldDeliverVerboseProgressDespiteSourceSuppression());
-    const hasVisibleRegularVerboseToolProgress = () =>
-      shouldEmitVerboseProgress() &&
-      !shouldEmitFullVerboseProgress() &&
-      shouldSendVerboseProgressMessages() &&
-      ctx.InboundEventKind !== "room_event" &&
-      !shouldSuppressProgressDelivery();
-    let observedVisibleToolErrorProgress = false;
-    const markVisibleToolErrorProgress = () => {
-      if (hasVisibleRegularVerboseToolProgress()) {
-        observedVisibleToolErrorProgress = true;
-      }
-    };
-    const hasFailedProgressStatus = (payload: {
-      phase?: string;
-      status?: string;
-      exitCode?: number | null;
-    }) =>
-      payload.phase === "error" ||
-      payload.status === "failed" ||
-      payload.status === "error" ||
-      (typeof payload.exitCode === "number" && payload.exitCode !== 0);
-    const shouldSuppressToolErrorWarnings = () => {
-      if (params.replyOptions?.suppressToolErrorWarnings !== undefined) {
-        return params.replyOptions.suppressToolErrorWarnings;
-      }
-      if (!shouldEmitVerboseProgress()) {
-        return false;
-      }
-      return observedVisibleToolErrorProgress ? true : undefined;
-    };
-    const suppressToolErrorWarnings =
-      params.replyOptions?.suppressToolErrorWarnings ??
-      (observedVisibleToolErrorProgress ? true : undefined);
-    const onToolResultFromReplyOptions = params.replyOptions?.onToolResult;
-    const onPlanUpdateFromReplyOptions = params.replyOptions?.onPlanUpdate;
-    const onApprovalEventFromReplyOptions = params.replyOptions?.onApprovalEvent;
-    const onPatchSummaryFromReplyOptions = params.replyOptions?.onPatchSummary;
-    const allowSuppressedSourceProgressCallbacks =
-      params.replyOptions?.allowProgressCallbacksWhenSourceDeliverySuppressed === true;
-    const isChannelOwnedToolResultProgressPayload = (payload: ReplyPayload) => {
-      const text = normalizeOptionalString(payload.text);
-      return Boolean(text?.startsWith("🛠️") || text?.startsWith("🔧"));
-    };
-    const shouldForwardToolResultProgressCallback = (
-      payload: ReplyPayload,
-      isFastModeAutoProgress: boolean,
-    ) => {
-      if (isFastModeAutoProgress) {
-        return shouldForwardProgressCallback({ forwardWhenSourceDeliverySuppressed: true });
-      }
-      if (
-        allowSuppressedSourceProgressCallbacks &&
-        isChannelOwnedToolResultProgressPayload(payload)
-      ) {
-        return shouldForwardProgressCallback({ forwardWhenSourceDeliverySuppressed: true });
-      }
-      return shouldSendToolSummaries() && shouldForwardProgressCallback();
-    };
-    const shouldAllowQuietChannelOwnedProgressCallbacks = (options?: {
-      allowWhenToolSummariesHidden?: boolean;
-      requiresToolSummaryVisibility?: boolean;
-    }) =>
-      options?.requiresToolSummaryVisibility === true &&
-      (params.replyOptions?.suppressDefaultToolProgressMessages === true ||
-        options.allowWhenToolSummariesHidden === true);
-    let hasPendingDirectBlockReplyDelivery = false;
-    const waitForPendingDirectBlockReplyDelivery = async (abortSignal?: AbortSignal) => {
-      if (!hasPendingDirectBlockReplyDelivery) {
-        return;
-      }
-      // Direct block replies are queued asynchronously so lightweight replies do
-      // not wait for dispatcher idle. Flush only before later tool/progress
-      // callbacks and final completion where external ordering is visible.
-      hasPendingDirectBlockReplyDelivery = false;
-      await waitForReplyDispatcherIdle(dispatcher, abortSignal);
-    };
-    const shouldForwardProgressCallback = (options?: {
-      allowWhenToolSummariesHidden?: boolean;
-      forwardWhenSourceDeliverySuppressed?: boolean;
-      requiresToolSummaryVisibility?: boolean;
-    }) => {
-      if (
-        options?.requiresToolSummaryVisibility === true &&
-        !shouldSendToolSummaries() &&
-        !shouldAllowQuietChannelOwnedProgressCallbacks(options)
-      ) {
-        return false;
-      }
-      return (
-        !suppressAutomaticSourceDelivery ||
-        (allowSuppressedSourceProgressCallbacks &&
-          !sendPolicyDenied &&
-          options?.forwardWhenSourceDeliverySuppressed === true)
-      );
-    };
-    const preserveProgressCallbackStartOrder =
-      params.replyOptions?.preserveProgressCallbackStartOrder === true;
-    let progressCallbackStartTail = Promise.resolve();
-    const reserveProgressCallbackStart = () => {
-      const previousStart = progressCallbackStartTail;
-      let releaseStart: (() => void) | undefined;
-      progressCallbackStartTail = new Promise<void>((resolve) => {
-        releaseStart = resolve;
-      });
-      return {
-        previousStart,
-        releaseStart: () => releaseStart?.(),
-      };
-    };
-    const wrapProgressCallback = <Args extends unknown[], Result extends false | void>(
-      callback: ((...args: Args) => Promise<Result> | Result) | undefined,
-      options?: {
-        allowWhenToolSummariesHidden?: boolean;
-        forwardWhenSourceDeliverySuppressed?: boolean;
-        requiresToolSummaryVisibility?: boolean;
-        onForward?: (...args: Args) => Promise<void> | void;
-        onVisible?: (...args: Args) => Promise<void> | void;
-        waitForDirectBlockReplyDelivery?: boolean;
-      },
-    ): ((...args: Args) => Promise<Result | undefined>) | undefined => {
-      if (!callback) {
-        return undefined;
-      }
-      const runProgressCallback = async (
-        args: Args,
-        noteCallbackStarted: () => void,
-      ): Promise<Result | undefined> => {
-        try {
-          if (isDispatchOperationAborted()) {
-            return undefined;
-          }
-          getDispatchReplyOperation()?.recordActivity();
-          markProgress();
-          if (options?.waitForDirectBlockReplyDelivery) {
-            await waitForPendingDirectBlockReplyDelivery(getDispatchAbortOperation()?.abortSignal);
-            if (isDispatchOperationAborted()) {
-              return undefined;
-            }
-          }
-          if (shouldForwardProgressCallback(options)) {
-            if (preserveProgressCallbackStartOrder && options?.onForward) {
-              await options.onForward(...args);
-            } else if (!preserveProgressCallbackStartOrder) {
-              // Preserve the historical microtask boundary for unflagged channels.
-              await options?.onForward?.(...args);
-            }
-            const callbackResult = callback(...args);
-            noteCallbackStarted();
-            const result = await callbackResult;
-            if (result === false) {
-              return result;
-            }
-            await options?.onVisible?.(...args);
-          }
-          return undefined;
-        } finally {
-          noteCallbackStarted();
-        }
-      };
-      return (...args: Args) => {
-        if (!preserveProgressCallbackStartOrder) {
-          return runProgressCallback(args, () => undefined);
-        }
-        // Reserve source order synchronously. Release after callback invocation, not completion,
-        // so async presentation work stays concurrent without letting later activity overtake it.
-        const start = reserveProgressCallbackStart();
-        return (async () => {
-          await start.previousStart;
-          return await runProgressCallback(args, start.releaseStart);
-        })();
-      };
-    };
-
-    // Snapshot verbose progress visibility for this run: commentary
-    // classification in the CLI runners is wired once at run start, so a
-    // mid-run verbose toggle cannot move inter-tool commentary between lanes.
-    const deliverStandaloneCommentaryProgress = shouldEmitVerboseProgress();
-    const itemEventForwardingOptions = {
-      forwardWhenSourceDeliverySuppressed: true,
-      requiresToolSummaryVisibility: true,
-    } as const;
-    const canForwardItemEvents =
-      Boolean(params.replyOptions?.onItemEvent) &&
-      shouldForwardProgressCallback(itemEventForwardingOptions);
-    const canForwardSuppressedSourceItemEvents =
-      suppressAutomaticSourceDelivery &&
-      allowSuppressedSourceProgressCallbacks &&
-      canForwardItemEvents;
-    const forwardItemEvent = canForwardItemEvents
-      ? wrapProgressCallback(params.replyOptions?.onItemEvent, {
-          ...itemEventForwardingOptions,
-          waitForDirectBlockReplyDelivery: true,
-          onForward: (payload) =>
-            preserveProgressCallbackStartOrder &&
-            deliverStandaloneCommentaryProgress &&
-            payload.kind === "preamble"
-              ? noteCommentaryProgress(payload)
-              : undefined,
-          onVisible: (payload) => {
-            if (hasFailedProgressStatus(payload)) {
-              markVisibleToolErrorProgress();
-            }
-          },
-        })
-      : undefined;
-    const canConsumeItemEvents = deliverStandaloneCommentaryProgress || canForwardItemEvents;
-    // Item-event presence gates CLI commentary classification downstream, so
-    // the handler exists exactly when verbose buffers it or a channel consumes it.
-    const onItemEvent = canConsumeItemEvents
-      ? async (payload: Parameters<NonNullable<GetReplyOptions["onItemEvent"]>>[0]) => {
-          if (isDispatchOperationAborted()) {
-            return;
-          }
-          if (!forwardItemEvent) {
-            // The wrapped forwarder marks progress itself when present.
-            markProgress();
-          }
-          if (
-            (!forwardItemEvent || !preserveProgressCallbackStartOrder) &&
-            deliverStandaloneCommentaryProgress &&
-            payload.kind === "preamble"
-          ) {
-            await noteCommentaryProgress(payload);
-          }
-          return await forwardItemEvent?.(payload);
-        }
-      : undefined;
-    // Let draft-rendering channels yield their ephemeral commentary lines while
-    // the durable verbose commentary lane is delivering the same content.
-    params.replyOptions?.onVerboseProgressVisibility?.(
-      () =>
-        deliverStandaloneCommentaryProgress &&
-        shouldSendVerboseProgressMessages() &&
-        !shouldSuppressProgressDelivery(),
-    );
-
-    const replyResolver =
-      params.replyResolver ??
-      (await traceReplyPhase("reply.load_reply_resolver", () => loadGetReplyFromConfigRuntime()))
-        .getReplyFromConfig;
-    // Channel runtimes can outlive a config reload. Resolve one live snapshot
-    // per turn so reply setup and dispatch callbacks share the same authority.
-    const runtimeReplyConfig = getRuntimeConfigSnapshot() ?? cfg;
-    const replyConfig = withFullRuntimeReplyConfig(
-      params.configOverride
-        ? (applyMergePatch(runtimeReplyConfig, params.configOverride) as OpenClawConfig)
-        : runtimeReplyConfig,
-    );
-    recordAgentDispatchStarted();
-    const replyResult = await runWithDispatchLifecycleAdmission(
-      async () =>
-        await runWithDispatchAbortSignal(
-          getDispatchAbortSignal(),
-          () =>
-            traceReplyPhase("reply.run_reply_resolver", () =>
-              replyResolver(
-                ctx,
-                {
-                  ...getReplyOptions(),
-                  sourceReplyDeliveryMode,
-                  sessionPromptSourceReplyDeliveryMode: sessionStableSourceReplyDeliveryMode,
-                  ...({
-                    onSessionMetadataChanges: notifySessionMetadataChanges,
-                    onSessionPrepared: notePreparedSession,
-                  } satisfies InternalReplyResolverOptions),
-                  onObservedReplyDelivery: markObservedReplyDelivery,
-                  suppressToolErrorWarnings,
-                  shouldSuppressToolErrorWarnings,
-                  typingPolicy: typing.typingPolicy,
-                  suppressTyping: typing.suppressTyping,
-                  onPartialReply: wrapProgressCallback(params.replyOptions?.onPartialReply),
-                  onReasoningStream: wrapProgressCallback(params.replyOptions?.onReasoningStream),
-                  streamReasoningInNonStreamModes:
-                    params.replyOptions?.streamReasoningInNonStreamModes,
-                  onReasoningEnd: wrapProgressCallback(params.replyOptions?.onReasoningEnd),
-                  onAssistantMessageStart: wrapProgressCallback(
-                    params.replyOptions?.onAssistantMessageStart,
-                  ),
-                  onBlockReplyQueued: wrapProgressCallback(params.replyOptions?.onBlockReplyQueued),
-                  onToolStart: wrapProgressCallback(params.replyOptions?.onToolStart, {
-                    allowWhenToolSummariesHidden:
-                      params.replyOptions?.allowToolLifecycleWhenProgressHidden === true,
-                    forwardWhenSourceDeliverySuppressed: true,
-                    requiresToolSummaryVisibility: true,
-                    waitForDirectBlockReplyDelivery: true,
-                    onForward: async () => {
-                      // Commentary precedes the tool that follows it.
-                      await flushPendingCommentaryProgress();
-                    },
-                  }),
-                  onItemEvent,
-                  commentaryProgressEnabled:
-                    deliverStandaloneCommentaryProgress ||
-                    canForwardSuppressedSourceItemEvents ||
-                    params.replyOptions?.commentaryProgressEnabled,
-                  reasoningPayloadsEnabled,
-                  commentaryPayloadsEnabled,
-                  onCommandOutput: wrapProgressCallback(params.replyOptions?.onCommandOutput, {
-                    forwardWhenSourceDeliverySuppressed: true,
-                    requiresToolSummaryVisibility: true,
-                    waitForDirectBlockReplyDelivery: true,
-                    onVisible: (payload) => {
-                      if (hasFailedProgressStatus(payload)) {
-                        markVisibleToolErrorProgress();
-                      }
-                    },
-                  }),
-                  onCompactionStart: wrapProgressCallback(params.replyOptions?.onCompactionStart, {
-                    allowWhenToolSummariesHidden:
-                      params.replyOptions?.allowToolLifecycleWhenProgressHidden === true,
-                    forwardWhenSourceDeliverySuppressed: true,
-                    requiresToolSummaryVisibility: true,
-                    waitForDirectBlockReplyDelivery: true,
-                  }),
-                  onCompactionEnd: wrapProgressCallback(params.replyOptions?.onCompactionEnd, {
-                    allowWhenToolSummariesHidden:
-                      params.replyOptions?.allowToolLifecycleWhenProgressHidden === true,
-                    forwardWhenSourceDeliverySuppressed: true,
-                    requiresToolSummaryVisibility: true,
-                    waitForDirectBlockReplyDelivery: true,
-                  }),
-                  onToolResult: (payload: ReplyPayload) => {
-                    getDispatchReplyOperation()?.recordActivity();
-                    markProgress();
-                    const run = async () => {
-                      if (isDispatchOperationAborted()) {
-                        return;
-                      }
-                      await waitForPendingDirectBlockReplyDelivery(
-                        getDispatchAbortOperation()?.abortSignal,
-                      );
-                      if (isDispatchOperationAborted()) {
-                        return;
-                      }
-                      markInboundDedupeReplayUnsafe();
-                      // Buffered commentary preceded this tool; land it before the summary.
-                      await flushPendingCommentaryProgress();
-                      // When the operator opts into messages.suppressToolErrors, never
-                      // surface tool-error tool-result payloads as channel progress,
-                      // regardless of source delivery mode. payloads.ts already drops
-                      // the warning text; this drops the visible progress delivery too.
-                      if (
-                        payload.isError === true &&
-                        replyConfig.messages?.suppressToolErrors === true
-                      ) {
-                        return;
-                      }
-                      const isFastModeAutoProgress = isFastModeAutoProgressPayload(payload);
-                      const isFastModeAutoProgressDelivery =
-                        isFastModeAutoProgress &&
-                        shouldDeliverFastModeAutoProgressDespiteSourceSuppression();
-                      const isForcedToolProgress =
-                        shouldDeliverForcedToolProgressDespiteSourceSuppression();
-                      const progressCallbackForwarded = shouldForwardToolResultProgressCallback(
-                        payload,
-                        isFastModeAutoProgress,
-                      );
-                      if (progressCallbackForwarded) {
-                        await onToolResultFromReplyOptions?.(payload);
-                      }
-                      if (isDispatchOperationAborted()) {
-                        return;
-                      }
-                      if (
-                        isFastModeAutoProgress &&
-                        progressCallbackForwarded &&
-                        onToolResultFromReplyOptions
-                      ) {
-                        return;
-                      }
-                      if (sendPolicyDenied) {
-                        return;
-                      }
-                      if (
-                        shouldSuppressProgressDelivery() &&
-                        !isFastModeAutoProgressDelivery &&
-                        !isForcedToolProgress &&
-                        !hasAskUserPayload(payload)
-                      ) {
-                        return;
-                      }
-                      const visibleToolPayload = isForcedToolProgress
-                        ? payload
-                        : resolveToolDeliveryPayload(payload);
-                      if (!visibleToolPayload) {
-                        return;
-                      }
-                      const ttsPayload = await maybeApplyTtsWithFinalizationLease({
-                        payload: visibleToolPayload,
-                        cfg,
-                        channel: deliveryChannel,
-                        kind: "tool",
-                        ttsAuto: sessionTtsAuto,
-                        agentId: sessionAgentId,
-                        accountId: replyRoute.accountId,
-                      });
-                      const normalizedPayload = await normalizeReplyMediaPayload(ttsPayload);
-                      const deliveryPayload = isForcedToolProgress
-                        ? normalizedPayload
-                        : resolveToolDeliveryPayload(normalizedPayload);
-                      if (!deliveryPayload) {
-                        return;
-                      }
-                      if (isDispatchOperationAborted()) {
-                        return;
-                      }
-                      if (
-                        shouldSuppressLateTextOnlyToolProgress(deliveryPayload) &&
-                        !isFastModeAutoProgressPayload(deliveryPayload) &&
-                        !isForcedToolProgress
-                      ) {
-                        return;
-                      }
-                      if (shouldSuppressMessageToolOnlyTextErrorProgress(deliveryPayload)) {
-                        return;
-                      }
-                      if (
-                        shouldSuppressDefaultToolProgressMessages() &&
-                        !isFastModeAutoProgressPayload(deliveryPayload) &&
-                        !isForcedToolProgress
-                      ) {
-                        const hasMedia =
-                          resolveSendableOutboundReplyParts(deliveryPayload).hasMedia;
-                        if (
-                          !hasMedia &&
-                          !hasExecApprovalPayload(deliveryPayload) &&
-                          !hasAskUserPayload(deliveryPayload)
-                        ) {
-                          return;
-                        }
-                      }
-                      if (deliveryPayload.isError === true) {
-                        markVisibleToolErrorProgress();
-                      }
-                      const askUserQuestionId = readAskUserQuestionId(deliveryPayload);
-                      if (
-                        askUserQuestionId !== undefined &&
-                        !(await isAskUserPromptPending(askUserQuestionId))
-                      ) {
-                        return;
-                      }
-                      if (isDispatchOperationAborted()) {
-                        return;
-                      }
-                      if (shouldRouteToOriginating) {
-                        await sendPayloadAsync(deliveryPayload, undefined, false);
-                      } else {
-                        markInboundDedupeReplayUnsafe();
-                        const delivered = dispatcher.sendToolResult(deliveryPayload);
-                        if (delivered && hasAskUserPayload(deliveryPayload)) {
-                          // ask_user blocks until this callback resolves; drain its prompt now
-                          // or the answerable UI can remain queued behind the blocked agent run.
-                          await waitForReplyDispatcherIdle(
-                            dispatcher,
-                            getDispatchAbortOperation()?.abortSignal,
-                          );
-                        }
-                      }
-                    };
-                    return run();
-                  },
-                  onPlanUpdate: async (payload) => {
-                    if (isDispatchOperationAborted()) {
-                      return;
-                    }
-                    const steps = normalizeAgentPlanSteps(payload.steps);
-                    const normalized = {
-                      phase: payload.phase,
-                      title: payload.title,
-                      explanation: payload.explanation,
-                      steps,
-                      source: payload.source,
-                    };
-                    markProgress();
-                    await waitForPendingDirectBlockReplyDelivery(
-                      getDispatchAbortOperation()?.abortSignal,
-                    );
-                    if (isDispatchOperationAborted()) {
-                      return;
-                    }
-                    markInboundDedupeReplayUnsafe();
-                    if (
-                      shouldForwardProgressCallback({
-                        forwardWhenSourceDeliverySuppressed: true,
-                        requiresToolSummaryVisibility: true,
-                      })
-                    ) {
-                      await onPlanUpdateFromReplyOptions?.(normalized);
-                    }
-                    if (isDispatchOperationAborted()) {
-                      return;
-                    }
-                    if (payload.phase !== "update" || shouldSuppressDefaultToolProgressMessages()) {
-                      return;
-                    }
-                    await sendPlanUpdate({
-                      explanation: normalized.explanation,
-                      steps,
-                    });
-                  },
-                  onApprovalEvent: async (payload) => {
-                    if (isDispatchOperationAborted()) {
-                      return;
-                    }
-                    markProgress();
-                    await waitForPendingDirectBlockReplyDelivery(
-                      getDispatchAbortOperation()?.abortSignal,
-                    );
-                    if (isDispatchOperationAborted()) {
-                      return;
-                    }
-                    markInboundDedupeReplayUnsafe();
-                    if (
-                      shouldForwardProgressCallback({
-                        forwardWhenSourceDeliverySuppressed: true,
-                        requiresToolSummaryVisibility: true,
-                      })
-                    ) {
-                      await onApprovalEventFromReplyOptions?.(payload);
-                    }
-                    if (isDispatchOperationAborted()) {
-                      return;
-                    }
-                    if (
-                      payload.phase !== "requested" ||
-                      shouldSuppressDefaultToolProgressMessages()
-                    ) {
-                      return;
-                    }
-                    const label = summarizeApprovalLabel({
-                      status: payload.status,
-                      command: payload.command,
-                      message: payload.message,
-                    });
-                    if (!label) {
-                      return;
-                    }
-                    await maybeSendWorkingStatus(label);
-                  },
-                  onPatchSummary: async (payload) => {
-                    if (isDispatchOperationAborted()) {
-                      return;
-                    }
-                    markProgress();
-                    await waitForPendingDirectBlockReplyDelivery(
-                      getDispatchAbortOperation()?.abortSignal,
-                    );
-                    if (isDispatchOperationAborted()) {
-                      return;
-                    }
-                    markInboundDedupeReplayUnsafe();
-                    if (
-                      shouldForwardProgressCallback({
-                        forwardWhenSourceDeliverySuppressed: true,
-                        requiresToolSummaryVisibility: true,
-                      })
-                    ) {
-                      await onPatchSummaryFromReplyOptions?.(payload);
-                    }
-                    if (isDispatchOperationAborted()) {
-                      return;
-                    }
-                    if (payload.phase !== "end" || shouldSuppressDefaultToolProgressMessages()) {
-                      return;
-                    }
-                    const label = summarizePatchLabel({
-                      summary: payload.summary,
-                      title: payload.title,
-                    });
-                    if (!label) {
-                      return;
-                    }
-                    await maybeSendWorkingStatus(label);
-                  },
-                  onBlockReply: (payload: ReplyPayload, context?: BlockReplyContext) => {
-                    markProgress();
-                    const run = async () => {
-                      if (isDispatchOperationAborted()) {
-                        return;
-                      }
-                      if (
-                        payload.isReasoning !== true &&
-                        payload.isCommentary !== true &&
-                        hasOutboundReplyContent(payload, { trimText: true })
-                      ) {
-                        markInboundDedupeReplayUnsafe();
-                      }
-                      // Buffered commentary preceded this block; deliver it first.
-                      await flushPendingCommentaryProgress();
-                      if (suppressDelivery) {
-                        return;
-                      }
-                      // Durable reasoning is a channel-owned lane; generic channels
-                      // keep the historical suppression unless they explicitly opt in.
-                      if (payload.isReasoning === true && !reasoningPayloadsEnabled) {
-                        return;
-                      }
-                      // Durable commentary is a channel-owned lane; generic channels keep the
-                      // historical suppression unless they explicitly opt in.
-                      if (payload.isCommentary === true && !commentaryPayloadsEnabled) {
-                        return;
-                      }
-                      // Accumulate block text for TTS generation after streaming.
-                      // Exclude status notices — they are informational UI signals
-                      // and must not be synthesised into the spoken reply. Display
-                      // lanes stay out too: they are presentation, never final text.
-                      const isStatusNotice = isReplyPayloadStatusNotice(payload);
-                      if (
-                        payload.text &&
-                        !isStatusNotice &&
-                        payload.isReasoning !== true &&
-                        payload.isCommentary !== true
-                      ) {
-                        const joinsBufferedTtsDirective =
-                          cleanBlockTtsDirectiveText?.hasBufferedDirectiveText() === true;
-                        if (accumulatedBlockText.length > 0) {
-                          accumulatedBlockText += "\n";
-                        }
-                        accumulatedBlockText += payload.text;
-                        if (accumulatedBlockTtsText.length > 0 && !joinsBufferedTtsDirective) {
-                          accumulatedBlockTtsText += "\n";
-                        }
-                        accumulatedBlockTtsText += payload.text;
-                        blockCount++;
-                      }
-                      const visiblePayload =
-                        payload.text &&
-                        cleanBlockTtsDirectiveText &&
-                        !isStatusNotice &&
-                        payload.isReasoning !== true &&
-                        payload.isCommentary !== true
-                          ? (() => {
-                              const text = cleanBlockTtsDirectiveText.push(payload.text);
-                              return copyReplyPayloadMetadata(payload, {
-                                ...payload,
-                                text: text.trim() ? text : undefined,
-                              });
-                            })()
-                          : payload;
-                      if (!hasOutboundReplyContent(visiblePayload, { trimText: true })) {
-                        return;
-                      }
-                      // Channels that keep a live draft preview may need to rotate their
-                      // preview state at the logical block boundary before queued block
-                      // delivery drains asynchronously through the dispatcher.
-                      const payloadMetadata = getReplyPayloadMetadata(payload);
-                      const queuedContext =
-                        payloadMetadata?.assistantMessageIndex !== undefined
-                          ? {
-                              ...context,
-                              assistantMessageIndex: payloadMetadata.assistantMessageIndex,
-                            }
-                          : context;
-                      if (!suppressAutomaticSourceDelivery) {
-                        await params.replyOptions?.onBlockReplyQueued?.(
-                          visiblePayload,
-                          queuedContext,
-                        );
-                      }
-                      if (isDispatchOperationAborted()) {
-                        return;
-                      }
-                      const ttsPayload =
-                        payload.isReasoning === true || payload.isCommentary === true
-                          ? visiblePayload
-                          : await maybeApplyTtsWithFinalizationLease({
-                              payload: visiblePayload,
-                              cfg,
-                              channel: deliveryChannel,
-                              kind: "block",
-                              ttsAuto: sessionTtsAuto,
-                              agentId: sessionAgentId,
-                              accountId: replyRoute.accountId,
-                            });
-                      const normalizedPayload = await normalizeReplyMediaPayload(ttsPayload);
-                      if (isDispatchOperationAborted()) {
-                        return;
-                      }
-                      if (shouldRouteToOriginating) {
-                        await sendPayloadAsync(
-                          normalizedPayload,
-                          context?.abortSignal,
-                          false,
-                          "block",
-                        );
-                      } else {
-                        markInboundDedupeReplayUnsafe();
-                        const delivered = dispatcher.sendBlockReply(normalizedPayload);
-                        if (delivered) {
-                          hasPendingDirectBlockReplyDelivery = true;
-                        }
-                      }
-                    };
-                    return run();
-                  },
-                },
-                replyConfig,
-              ),
-            ),
-          trackDispatchLifecycleWork,
-        ),
-    );
-    const sessionMetadataChanges = takeCommandSessionMetadataChanges(ctx);
-    notifySessionMetadataChanges(sessionMetadataChanges);
-    const finalDispatchAcquisition = await ensureDispatchReplyOperation("dispatch");
-    if (finalDispatchAcquisition.status === "aborted") {
-      return finishReplyOperationAbortedDispatch();
-    }
-    if (finalDispatchAcquisition.status === "busy") {
-      return finishReplyOperationBusyDispatch({
-        recordAgentDispatchCompleted: true,
-        ...(sessionMetadataChangesForResult
-          ? { sessionMetadataChanges: sessionMetadataChangesForResult }
-          : {}),
-      });
-    }
-
-    if (ctx.AcpDispatchTailAfterReset === true) {
-      // Command handling prepared a trailing prompt after ACP in-place reset.
-      // Route that tail through ACP now (same turn) instead of embedded dispatch.
-      ctx.AcpDispatchTailAfterReset = false;
-      if (hookRunner?.hasHooks("reply_dispatch")) {
-        const tailDispatchResult = await runWithDispatchLifecycleAdmission(
-          async () =>
-            await runWithDispatchAbortSignal(
-              getDispatchAbortSignal(),
-              () =>
-                hookRunner.runReplyDispatch(
-                  createReplyDispatchEvent({
-                    ctx,
-                    runId: params.replyOptions?.runId,
-                    sessionKey: acpDispatchSessionKey,
-                    toolsAllow: params.replyOptions?.toolsAllow,
-                    images: params.replyOptions?.images,
-                    inboundAudio,
-                    sessionTtsAuto,
-                    ttsChannel: deliveryChannel,
-                    suppressUserDelivery: suppressHookUserDelivery,
-                    suppressReplyLifecycle: suppressHookReplyLifecycle,
-                    sourceReplyDeliveryMode,
-                    shouldRouteToOriginating,
-                    originatingChannel: routeReplyChannel,
-                    originatingTo: routeReplyTo,
-                    originatingAccountId: replyContextAccountId,
-                    originatingThreadId: routeReplyThreadId,
-                    originatingChatType: replyRoute.chatType,
-                    shouldSendToolSummaries,
-                    sendPolicy,
-                    isTailDispatch: true,
-                  }),
-                  {
-                    cfg,
-                    dispatcher: dispatchHookDispatcher,
-                    abortSignal: getPreDispatchAbortSignal() ?? params.replyOptions?.abortSignal,
-                    onReplyStart: params.replyOptions?.onReplyStart,
-                    recordProcessed,
-                    markIdle,
-                  },
-                ),
-              trackDispatchLifecycleWork,
-            ),
-        );
-        if (tailDispatchResult?.handled) {
-          recordAgentDispatchCompleted("completed");
-          completeDispatchReplyOperation();
-          return attachSourceReplyDeliveryMode({
-            queuedFinal: tailDispatchResult.queuedFinal,
-            counts: tailDispatchResult.counts,
-            ...(sessionMetadataChangesForResult
-              ? { sessionMetadataChanges: sessionMetadataChangesForResult }
-              : {}),
-          });
-        }
-      }
-    }
-
-    const replies = replyResult ? (Array.isArray(replyResult) ? replyResult : [replyResult]) : [];
-    const pendingFinalDelivery = {
-      storePath: sessionStoreEntry.storePath,
-      sessionKey: sessionStoreEntry.sessionKey ?? sessionKey,
-    };
-    const replyPendingIntentIds = new Set(
-      replies
-        .map((reply) => getReplyPayloadMetadata(reply)?.pendingFinalDeliveryIntentId)
-        .filter((intentId): intentId is string => Boolean(intentId)),
-    );
-    const pendingFinalDeliveryIdentity = capturePendingFinalDeliveryIdentity({
-      ...pendingFinalDelivery,
-      intentId: replyPendingIntentIds.size === 1 ? [...replyPendingIntentIds][0] : undefined,
-    });
-    // Final delivery is outside the progress wrappers. Wait until every source-ordered callback
-    // has at least started so a delayed tool/reasoning transition cannot appear after the final.
-    if (preserveProgressCallbackStartOrder) {
-      await progressCallbackStartTail;
-    }
-    // Backstop: silent/streaming-delivered turns end without a visible final
-    // reply; trailing commentary must still land.
-    await flushPendingCommentaryProgress();
-    const beforeAgentRunBlocked = replies.some(
-      (reply) => getReplyPayloadMetadata(reply)?.beforeAgentRunBlocked === true,
-    );
-
-    let queuedFinal = false;
-    let routedFinalCount = 0;
-    let attemptedFinalDelivery = false;
-    let finalDeliveryFailed = false;
-    const finalDeliveries: Array<{
-      outcome: Promise<ReplyDispatchDeliveryOutcome>;
-      payload: ReplyPayload;
-    }> = [];
-    let allQueuedFinalsObserved = true;
-    // Explicit command turns (native or authorized text-slash like /compact) are
-    // user-initiated, so a marked terminal reply for the command bypasses
-    // room_event suppression. Ambient marked notices (no CommandTurn) stay
-    // suppressed in room_event. sendPolicy: deny still suppresses everything.
-    // Uses the same helper as the source-reply visibility policy so the bypass
-    // and the policy stay aligned.
-    const shouldDeliverDespiteSourceReplySuppression = (reply: ReplyPayload) =>
-      suppressAutomaticSourceDelivery &&
-      !sendPolicyDenied &&
-      getReplyPayloadMetadata(reply)?.deliverDespiteSourceReplySuppression === true &&
-      (ctx.InboundEventKind !== "room_event" || explicitCommandTurnCtx);
-    const sentFinalPayloadDedupeKeys = new Set<string>();
-    for (const [replyIndex, reply] of replies.entries()) {
-      throwIfDispatchOperationAborted();
-      // Durable reasoning is a channel-owned lane; generic channels keep the
-      // historical suppression unless they explicitly opt in.
-      if (reply.isReasoning === true && !reasoningPayloadsEnabled) {
-        continue;
-      }
-      if (reply.isCommentary === true && !commentaryPayloadsEnabled) {
-        continue;
-      }
-      if (suppressDelivery && !shouldDeliverDespiteSourceReplySuppression(reply)) {
-        if (hasOutboundReplyContent(reply, { trimText: true })) {
-          logVerbose(
-            [
-              `dispatch-from-config: final reply suppressed by ${deliverySuppressionReason || "source delivery policy"}`,
-              `(session=${acpDispatchSessionKey ?? sessionKey ?? "unknown"}`,
-              `provider=${ctx.Provider ?? "unknown"}`,
-              `surface=${ctx.Surface ?? "unknown"}`,
-              `chatType=${chatType ?? "unknown"}`,
-              `inboundEventKind=${ctx.InboundEventKind ?? "unknown"}`,
-              `message=${ctx.MessageSidFull ?? ctx.MessageSid ?? "unknown"}`,
-              `${formatSuppressedReplyPayloadForLog(reply)})`,
-            ].join(" "),
-          );
-        }
-        continue;
-      }
-      const finalPayloadDedupeKey = createFinalDispatchPayloadDedupeKey(reply);
-      if (sentFinalPayloadDedupeKeys.has(finalPayloadDedupeKey)) {
-        continue;
-      }
-      sentFinalPayloadDedupeKeys.add(finalPayloadDedupeKey);
-      attemptedFinalDelivery = true;
-      const finalReply = await sendFinalPayload(reply, { deliveryId: String(replyIndex) });
-      queuedFinal = finalReply.queuedFinal || queuedFinal;
-      routedFinalCount += finalReply.routedFinalCount;
-      if (finalReply.queuedFinal) {
-        if (finalReply.dispatcherOutcome) {
-          finalDeliveries.push({ outcome: finalReply.dispatcherOutcome, payload: reply });
-        } else {
-          allQueuedFinalsObserved = false;
-        }
-      }
-      if (!finalReply.queuedFinal && finalReply.routedFinalCount === 0) {
-        finalDeliveryFailed = true;
-      }
-    }
-
-    if (attemptedFinalDelivery && !finalDeliveryFailed) {
-      if (queuedFinal && allQueuedFinalsObserved) {
-        // Delivery observers run from the queue itself, so direct low-level callers
-        // reconcile too; the settle task only makes lifecycle owners await it.
-        const reconcilePendingFinal = Promise.all(
-          finalDeliveries.map(async (delivery) => ({
-            outcome: await delivery.outcome,
-            payload: delivery.payload,
-          })),
-        )
-          .then(async (deliveries) => {
-            await reconcilePendingFinalDeliveryAfterSettlement({
-              ...pendingFinalDelivery,
-              deliveries,
-              identity: pendingFinalDeliveryIdentity,
-              replies,
-            });
-          })
-          .catch((error: unknown) => {
-            logVerbose(
-              `dispatch-from-config: pending final reconciliation failed: ${formatErrorMessage(error)}`,
-            );
-          });
-        registerReplyDispatcherSettledTask(dispatcher, () => reconcilePendingFinal);
-      } else {
-        // Routed delivery has a transport result already. Custom dispatchers that
-        // do not expose the core observer retain the legacy queue-admission behavior.
-        await clearPendingFinalDeliveryAfterSuccess({
-          ...pendingFinalDelivery,
-          identity: pendingFinalDeliveryIdentity,
-        });
-      }
-      // Register successful queued cleanup before honoring a late abort. The
-      // outer settle owner still runs it from finally (#89115).
-      throwIfDispatchOperationAborted();
-    }
-
-    if (!suppressDelivery) {
-      const ttsMode = resolveConfiguredTtsMode(cfg, {
-        agentId: sessionAgentId,
-        channelId: deliveryChannel,
-        accountId: replyRoute.accountId,
-      });
-      // Generate TTS-only reply after block streaming completes (when there's no final reply).
-      // This handles the case where block streaming succeeds and drops final payloads,
-      // but we still want TTS audio to be generated from the accumulated block content.
-      if (
-        ttsMode === "final" &&
-        replies.length === 0 &&
-        blockCount > 0 &&
-        accumulatedBlockTtsText.trim()
-      ) {
-        try {
-          await waitForPendingDirectBlockReplyDelivery(getDispatchAbortSignal());
-          throwIfDispatchOperationAborted();
-          const ttsSyntheticReply = await maybeApplyTtsWithFinalizationLease({
-            payload: { text: accumulatedBlockTtsText },
-            cfg,
-            channel: deliveryChannel,
-            kind: "final",
-            ttsAuto: sessionTtsAuto,
-            agentId: sessionAgentId,
-            accountId: replyRoute.accountId,
-          });
-          throwIfDispatchOperationAborted();
-          // Only send if TTS was actually applied (mediaUrl exists)
-          if (ttsSyntheticReply.mediaUrl) {
-            // Send TTS-only payload (no text, just audio) so it doesn't duplicate the block content.
-            // Keep the spoken text only for hooks/archive consumers.
-            const ttsOnlyPayload = markReplyPayloadAsTtsSupplement(
-              {
-                mediaUrl: ttsSyntheticReply.mediaUrl,
-                audioAsVoice: ttsSyntheticReply.audioAsVoice,
-                spokenText: accumulatedBlockTtsText,
-                trustedLocalMedia: true,
-              },
-              accumulatedBlockTtsText,
-              { visibleTextAlreadyDelivered: true },
-            );
-            const normalizedTtsOnlyPayload = await normalizeReplyMediaPayload(ttsOnlyPayload);
-            throwIfDispatchOperationAborted();
-            const result = await routeReplyToOriginating(normalizedTtsOnlyPayload, {
-              abortSignal: getDispatchAbortSignal(),
-              kind: "final",
-            });
-            if (result) {
-              queuedFinal = result.ok || queuedFinal;
-              if (isRoutedReplyDelivered(result)) {
-                routedFinalCount += 1;
-              }
-              if (!result.ok) {
-                logVerbose(
-                  `dispatch-from-config: route-reply (tts-only) failed: ${result.error ?? "unknown error"}`,
-                );
-              }
-            } else {
-              throwIfDispatchOperationAborted();
-              markInboundDedupeReplayUnsafe();
-              const didQueue = dispatcher.sendFinalReply(normalizedTtsOnlyPayload);
-              queuedFinal = didQueue || queuedFinal;
-            }
-          }
-        } catch (err) {
-          if (isDispatchReplyOperationAbortedError(err)) {
-            throw err;
-          }
-          logVerbose(
-            `dispatch-from-config: accumulated block TTS failed: ${formatErrorMessage(err)}`,
-          );
-        }
-      }
-    }
-
-    await waitForPendingDirectBlockReplyDelivery(getDispatchAbortSignal());
-    const counts = dispatcher.getQueuedCounts();
-    counts.final += routedFinalCount;
-    commitInboundDedupeIfClaimed();
-    recordAgentDispatchCompleted("completed");
-    recordProcessed(
-      "completed",
-      pluginFallbackReason ? { reason: pluginFallbackReason } : undefined,
-    );
-    markIdle("message_completed");
-    completeDispatchReplyOperation();
-    return attachSourceReplyDeliveryMode({
-      queuedFinal,
-      counts,
-      ...(sessionMetadataChangesForResult
-        ? { sessionMetadataChanges: sessionMetadataChangesForResult }
-        : {}),
-      ...(getObservedReplyDelivery() ? { observedReplyDelivery: true } : {}),
-      ...(!queuedFinal && !getObservedReplyDelivery() && !emptyFinalAllowedAsSilent
-        ? { noVisibleReplyFallbackEligible: true }
-        : {}),
-      ...(beforeAgentRunBlocked ? { beforeAgentRunBlocked } : {}),
-    });
+    const finalized = await finalizeDispatchAndAudit(executed.state);
+    return finalized.result;
   } catch (err) {
+    const {
+      failDispatchReplyOperation,
+      finishReplyOperationAbortedDispatch,
+      inboundDedupeClaim,
+      markIdle,
+      recordAgentDispatchCompleted,
+      recordProcessed,
+    } = errorState;
     if (isDispatchReplyOperationAbortedError(err)) {
       return finishReplyOperationAbortedDispatch();
     }
     if (inboundDedupeClaim.status === "claimed") {
-      if (inboundDedupeReplayUnsafe) {
+      if (errorState.inboundDedupeReplayUnsafe) {
         commitInboundDedupe(inboundDedupeClaim.key);
       } else {
         releaseInboundDedupe(inboundDedupeClaim.key);
@@ -3058,4 +96,3 @@ async function dispatchReplyFromConfigInner(
     throw err;
   }
 }
-/* oxlint-disable max-lines -- TODO: split this grandfathered oversized file. */
